% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{across_an1}
\alias{across_an1}
\title{across an1 dummy}
\usage{
across_an1(across, across_subset)
}
\arguments{
\item{across}{Character value or NULL. If character, the name of the grouping variable
that contains the group assignment across which analysis should
be conducted. If NULL, no grouping is done.}

\item{across.subset, across_subset}{Character vector or NULL. If NULL, all groups
are considered. If character, specifies the groups of the
grouping variable that are of interest. Prefixing group names with
\emph{'-'} excludes them.}
}
\description{
across an1 dummy
}
