% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{across_an2}
\alias{across_an2}
\title{across an2 dummy}
\usage{
across_an2(across, across_subset)
}
\arguments{
\item{across}{Character vector or NULL. If character, the name of the grouping variables
that contain the group assignment across which analysis should
be conducted. If NULL, no grouping is done.}

\item{across.subset, across_subset}{Character vector or NULL. If NULL, all groups
are considered. If character, specifies the groups of the
grouping variable that are of interest. Prefixing group names with
\emph{'-'} excludes them.}
}
\description{
across an2 dummy
}
