% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame-handling.R
\name{across_options}
\alias{across_options}
\title{Across input}
\usage{
across_options(df, n.across.subset = 25)
}
\arguments{
\item{df}{A data.frame}

\item{n.across.subset}{Numeric value. Denotes the maximal number of unique
groups the across options to be returned may contain.}
}
\description{
Returns input options for the \code{across} argument.
}
