% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-wrapper-shinyWidgets.R
\name{across_picker_input}
\alias{across_picker_input}
\alias{across_subset_picker_input}
\alias{colorpanel_picker_input}
\alias{color_picker_input}
\title{Picker inputs}
\usage{
across_picker_input(ns, id = "across", choices, selected = "condition", ...)

across_subset_picker_input(
  ns,
  id = "across_subset",
  choices,
  selected,
  multiple = TRUE,
  ...
)

colorpanel_picker_input(ns, id = "pt_clrp")

color_picker_input(ns, id = "pt_color")
}
\description{
Picker inputs
}
