% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{across_vis1}
\alias{across_vis1}
\title{across vis1 dummy}
\usage{
across_vis1(across, across_subset)
}
\arguments{
\item{across}{Character value or NULL. If character, specifies the name of the grouping variable
that contains the group assignment across which results are displayed. If NULL,
no grouping is done.}

\item{acrosss.subset, across_subset}{Character vector or NULL. If character, specifies groups of the
grouping variable that are of interest. Prefixing group names with \emph{'-'} excludes them.
If NULL, all groups are considered.}
}
\description{
across vis1 dummy
}
