% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{across_vis2}
\alias{across_vis2}
\title{across vis2 dummy}
\usage{
across_vis2(across, across_subset)
}
\arguments{
\item{across}{Character vector or NULL. If character, the names of the grouping variables
that contain the group assignments across which results are displayed. If NULL, no grouping
is done.}

\item{across_subset}{List of character vectors. Must be named according
to the input of argument \code{across}. Named slots of the list should be character
vectors containing the groups of interest. Prefixing group names with
\emph{'-'} excludes them.}
}
\description{
across vis2 dummy
}
