% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{adapt_reference}
\alias{adapt_reference}
\title{Adapt glue reference}
\usage{
adapt_reference(input, sg, pl = NULL, zero = "")
}
\arguments{
\item{input}{Vector to be checked.}

\item{sg}{Character value to return if length of \code{input} is 1.}

\item{pl}{Character value to return if length of \code{input} is > 1.
If set to NULL an \emph{'s'} is attached to in put of \code{sg}.}

\item{zero}{Character value to treturn if lengt of \code{input} is 0.}
}
\value{
Either sg or pl.
}
\description{
Switch between plural or singular reference.
}
