% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Analysis.R
\name{addClusterVarsKmeans}
\alias{addClusterVarsKmeans}
\alias{addClusterVarsKmeans,Analysis-method}
\title{Add kmeans cluster variables}
\usage{
addClusterVarsKmeans(object, ...)

\S4method{addClusterVarsKmeans}{Analysis}(
  object,
  ks,
  methods_kmeans = "Hartigan-Wong",
  prefix = "",
  naming = "{method_kmeans}_k{k}",
  overwrite = FALSE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{ks}{Numeric vector. Denotes all options for k-clusters. Values <1 are discarded.
Is converted into an integer vector.}
}
\description{
Adds kmeans cluster results in form of grouping variables
to the data of the input object.
}
\details{
Iterates over all possible combinations of the arguments \code{ks}
and \code{methods_kmeans} and adds the corresponding grouping
variables to the objects data. Cluster variables are named corresponding to the input
of argument \code{naming}.
}
