% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Analysis.R
\name{addClusterVarsPam}
\alias{addClusterVarsPam}
\alias{addClusterVarsPam,Analysis-method}
\title{Add PAM cluster variables}
\usage{
addClusterVarsPam(object, ...)

\S4method{addClusterVarsPam}{Analysis}(
  object,
  ks,
  methods_pam = "euclidean",
  prefix = "",
  naming = "{method_pam}_k{k}",
  overwrite = FALSE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{ks}{Numeric vector. Denotes all options for k-clusters. Values <1 are discarded.
Is converted into an integer vector.}
}
\description{
Adds PAM cluster results in form of grouping variables
to the data of the input object.
}
\details{
Iterates over all possible combinations of the arguments \code{ks} and
\code{methods_pam} and adds the corresponding grouping
variables to the objects data. Cluster variables are named corresponding to the input
of arguments \code{naming}.
}
