% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Analysis.R
\name{addGroupingVars}
\alias{addGroupingVars}
\alias{addGroupingVars,Analysis-method}
\title{Add grouping variables}
\usage{
addGroupingVars(object, ...)

\S4method{addGroupingVars}{Analysis}(object, grouping_df)
}
\arguments{
\item{grouping_df}{Data.frame that contains the grouping variables.}
}
\description{
Adds grouping variables to the @data of the object.
}
