% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-Correlation.R
\name{adjust_corr_mtr}
\alias{adjust_corr_mtr}
\title{Adjust correlation matrix}
\usage{
adjust_corr_mtr(mtr, type = "lower", diagonal = TRUE)
}
\arguments{
\item{mtr}{Correlation or p-value matrix.}

\item{type}{Character value. Denotes the adjustment type. Either of \emph{'lower',
'upper'} or \emph{'complete'}.}

\item{diagonal}{Logical. Indicates whether the diagonal of the correlation
matrix should be kept or should be converted to NA.}
}
\value{
Adjusted input matrix.
}
\description{

}
