% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Clustering.R,
%   R/S4-Analysis.R
\name{agglomerateHierarchicalTrees}
\alias{agglomerateHierarchicalTrees}
\alias{agglomerateHierarchicalTrees,Clustering-method}
\alias{agglomerateHierarchicalTrees,Analysis-method}
\title{Agglomerate hierarchical trees}
\usage{
agglomerateHierarchicalTrees(object, ...)

\S4method{agglomerateHierarchicalTrees}{Clustering}(
  object,
  methods_dist = "euclidean",
  methods_aggl = "ward.D",
  verbose = TRUE,
  ...
)

\S4method{agglomerateHierarchicalTrees}{Analysis}(
  object,
  methods_dist = methods_dist,
  methods_aggl = methods_aggl,
  verbose = verbose,
  ...
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{...}{Additional arguments given to \code{stats::hclust()}.}

\item{methods_dist}{Character vector. Specifies the distance methods
of interest. Use \code{validMethodsDist()} to obtain all valid input options.}

\item{methods_aggl}{Character vector. Specifies the agglomerative
methods of interest. Use \code{validMethodsAggl()} to obtain all valid input options.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object.
}
\description{
Agglomerates hierarchical trees from a list
of distance matrices.
}
