% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes-and-spectra.R
\name{all_color_palettes}
\alias{all_color_palettes}
\alias{all_color_palettes_vec}
\alias{all_color_spectra}
\alias{all_color_spectra_vec}
\title{Color palette names}
\usage{
all_color_palettes()

all_color_palettes_vec()

all_color_spectra()

all_color_spectra_vec()
}
\value{
A named list.
}
\description{
Returns all currently valid color palettes or -spectra.
}
