% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{all_colorspectra}
\alias{all_colorspectra}
\title{Deprecated}
\usage{
all_colorspectra()
}
\value{

}
\description{
\code{all_colorspectra()} is deprecated due to naming issues. Use
\code{all_color_spectra()} instead.
}
