% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-tests.R
\name{all_vars_discrete}
\alias{all_vars_discrete}
\title{Check data.frame variables}
\usage{
all_vars_discrete(df)
}
\arguments{
\item{df}{A data.frame.}
}
\value{
Logical value.
}
\description{
Check data.frame variables
}
