% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-AnalysisAspect.R
\docType{class}
\name{AnalysisAspect-class}
\alias{AnalysisAspect-class}
\alias{AnalysisAspect}
\title{The \code{AnalysisAspect}-class}
\description{
S4-class that provides the basic slots for any analysis
aspect such as clustering, dimensional reduction, outlier detection etc.
}
\section{Slots}{

\describe{
\item{\code{data}}{data.frame. The data on which the analysis bases on.}

\item{\code{key_name}}{character. The name of the variable that is used to identify
each observation uniquely.}

\item{\code{meta}}{data.frame. Data that was part of the input data but is not supposed
to be included in analysis steps.}

\item{\code{methods}}{list. A list of objects of S4-classes depending on the analysis
aspect.}

\item{\code{variables_grouping}}{character. The names of all grouping variables
of the input data - variables of class character or factor. (Does not include
variable of slot @key_name)}

\item{\code{variables_logical}}{character. The names of all logical variables of
the input data.}

\item{\code{variables_numeric}}{character. The names of all numeric variables
based on which outlier detection is conducted.}
}}

