% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{argument_dummy}
\alias{argument_dummy}
\title{Argument dummy}
\usage{
argument_dummy(df, normalize, verbose)
}
\arguments{
\item{df}{A data.frame.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{across}{Character value. Denotes the discrete variable in the data.frame
across which the variables of interest are to be analyzed or displayed.}

\item{across.subset}{Character vector. The groups of interest that the grouping variable
denoted in \code{across} contains.

If there are more groups you are interested in than they are groups you are not interested
in specify those that you are not interested in prefixed with an \emph{'-'}.
Variables prefixed that way are discarded and the remaining are kept.}

\item{display.facets}{Logical value. If set to TRUE the plot is split via
\code{ggplot2::facet_wrap()} such that each variable gets it's own subplot.}

\item{pt.alpha}{Numeric value. Denotes the transperancy of the points.}

\item{pt.color}{Character value. Denotes the color with
which the points are displayed .}

\item{pt.num}{Numeric value. Regulates the number of points (sample size)
to prevent overplotting.}

\item{pt.shape}{Numeric or character value. If numeric, the respective
shape is taken for all points. If character, the respective variable is
mapped to the shape-aesthetic.}

\item{pt.size}{Numeric values. Denote size of the points.}

\item{ref.group}{Character value. Denotes the reference group for the statistical tests. Must
be one value of the variable specified in \code{across}.}

\item{relevel}{Logical value. If set to TRUE the input of \code{across.subset}
determines the new order in which the results are displayed.}

\item{scales, nrow, ncol}{Given to \code{ggplot2::facet_wrap()}.}

\item{test.pairwise}{Character value or NULL. If character, one of \emph{'none', 't.test', 'wilcox.test'}.}

\item{test.groupwise}{Character value or NULL. If character, one of \emph{'none', 'anova', 'kruskal.test'}.}

\item{valid.classes}{Character vector to specify the classes the input
might have in order not to be discarded.}

\item{variables}{Character vector. Denotes the variables
of interest. If set to NULL all valid variables of input for \code{df} are
considered (as long as the total number does not exceeds the limit).

If there are more variables you are interested in than they are variables
you are not interested in specify those that you are not interested in prefixed
with an \emph{'-'}. Variables prefixed that way are discarded and the remaining are kept.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Argument dummy
}
