% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame-handling.R
\name{arrange_rows}
\alias{arrange_rows}
\title{Arrange rows}
\usage{
arrange_rows(df, across, verbose)
}
\arguments{
\item{df}{A data.frame with numeric variables that.}

\item{across}{Character. Either \emph{'max} or \emph{'min'}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An arranged data.frame.
}
\description{
Arranges the rows of a data.frame according to
the positions of their variable-maximum or -minimum.
}
