% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{assign_obj}
\alias{assign_obj}
\title{Assign objects into the global environment}
\usage{
assign_obj(assign, object, name)
}
\arguments{
\item{assign}{Logical.}

\item{object}{The object to be assigned.}

\item{name}{The name of the assigned object.}
}
\description{
Assign objects into the global environment
}
