% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-default.R
\name{assignDefault}
\alias{assignDefault}
\title{S4 Default Assignment}
\usage{
assignDefault(object, default_slot = "default")
}
\arguments{
\item{object}{The S4 object that has a slot @default which is
a named list.}

\item{default_slot}{Character value. The name of the slot that contains
the default list.}
}
\value{

}
\description{
Extracts the default list from the respective slot
and assigns the content of the list's slot to the environment of
the calling function if the argument input was \code{NULL}.
}
