% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call-flexibly.R
\name{call_flexibly}
\alias{call_flexibly}
\title{Construct a flexible function call}
\usage{
call_flexibly(
  fn.name,
  fn.namespace,
  default_list = list(),
  verbose = TRUE,
  value.fail = NULL,
  value.skip = NULL
)
}
\arguments{
\item{fn.name}{Character value. Denotes the function to be called.}

\item{fn.namespace}{Character value. Denotes the namespace/package from which to call \emph{fn.name}.}

\item{default_list}{A named list of arguments that can not be specified by the user.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{value.fail}{The return value in case the function call results in an error.}

\item{value.skip}{The return value in case of \code{fn.name} is specified as FALSE by the user..}
}
\value{
The return value of \emph{fn.name}.
}
\description{
Programming aid: See details for more information on how to use it.
}
\details{
This function takes two strings as input denoting
the function to be called and the namespace from which it is to be called. It expects
an object of the same name as \emph{fn.name} to be in it's calling environment specified with an identically named
argument from the user of the function from which \code{call_flexibly()} is called.

If that object is a list all named elements of that list are considered to be arguments with
which function \emph{fn.name} is to be called (= specified list). The names of that list are compared to the names
of \code{default_list}. Arguments specified in the default list can not be altered and are discarded from
the specified list with an informative warning. Subsequently the names of all remaining arguments are compared
to the valid arguments of the function to be called and discarded if unused arguments appear in order to
prevent the function call from failing. (This does not happen if the function to be called uses the dot-product '...').

If that object is a single TRUE the \code{fn.name} is called with \code{default_list} as input.

If that object is anything else the function call is skipped.
}
\examples{
# Not run:

 example_fun <- function(plot, runif){

   call_flexibly(fn.name = "plot",
                 fn.namespace = "base",
                 default_list = list(x = 1:10),
                 value.fail = "This failed.",
                 value.skip = "Okey, I skip that."
                 )

   call_flexibly(fn.name = "runif",
                 fn.namespace = "stats",
                 default_list = list(n = 100),
                 value.fail = "This failed.",
                 value.skip = 1:100)

 }


 # call the function

 example_fun(plot = list(y = 1:10, cex = 5, col = "red"), runif = list(max = 100, min = 1))

}
