% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous-data-frame.R
\name{check_across_subset}
\alias{check_across_subset}
\title{Subset the across-variables}
\usage{
check_across_subset(df, across, across.subset)
}
\arguments{
\item{df}{A data.frame that contains the variable specified in \code{across}.}

\item{across}{Character value. Denotes the discrete variable in the data.frame
across which the variables of interest are to be analyzed or displayed. Valid input
options can be obtained via \code{getAcrossOptions()}.}
}
\value{
A filtered data.frame, informative messages or an error.
}
\description{
Checks across and across.subset input and if at least one
of the across.subset values exists filters the data.frame accordingly.
}
