% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-check.R
\name{check_across_subset_negate}
\alias{check_across_subset_negate}
\title{Check across subset input}
\usage{
check_across_subset_negate(across, across.subset, all.groups)
}
\description{
Check across subset input
}
