% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-wrapper.R
\name{check_corr_availability}
\alias{check_corr_availability}
\title{Title}
\usage{
check_corr_availability(input, method.corr, across = NULL, fdb.fn = "stop")
}
\arguments{
\item{fdb.fn}{}
}
\value{

}
\description{
Title
}
