% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-check.R
\name{check_df_variables}
\alias{check_df_variables}
\title{Data.frame variable check}
\usage{
check_df_variables(
  df,
  valid.classes,
  variables = NULL,
  keep = NULL,
  verbose = TRUE
)
}
\arguments{
\item{df}{A data.frame.}

\item{valid.classes}{Character vector to specify the classes the input
might have in order not to be discarded.}

\item{variables}{Character vector. Denotes the variables
of interest. If set to NULL all valid variables of input for \code{df} are
considered (as long as the total number does not exceeds the limit).

If there are more variables you are interested in than they are variables
you are not interested in specify those that you are not interested in prefixed
with an \emph{'-'}. Variables prefixed that way are discarded and the remaining are kept.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input \code{df} with all selected variables.
}
\description{
Selects the variables denoted in
\code{keep} and \code{variables}. The letter ones
are checked for validity.
}
