% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-check.R
\name{check_directories}
\alias{check_directories}
\title{Check directory input}
\usage{
check_directories(directories, ref)
}
\arguments{
\item{directories}{Character vector. Directories to check.}
}
\value{
An informative error message or an invisible TRUE.
}
\description{
Check directory input
}
