% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust-wrapper.R
\name{check_hclust_availability}
\alias{check_hclust_availability}
\title{Title}
\usage{
check_hclust_availability(input, method.dist, method.aggl, fdb.fn = "stop")
}
\arguments{
\item{fdb.fn}{}
}
\value{

}
\description{
Title
}
