% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust-wrapper.R
\name{check_hclust_methods}
\alias{check_hclust_methods}
\title{Title}
\usage{
check_hclust_methods(
  method.dist = NULL,
  method.aggl = NULL,
  methods.dist = NULL,
  methods.aggl = NULL
)
}
\arguments{
\item{methods.aggl}{}
}
\value{

}
\description{
Title
}
