% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans-wrapper.R
\name{check_kmeans_availability}
\alias{check_kmeans_availability}
\title{Title}
\usage{
check_kmeans_availability(input, method.kmeans, centers, fdb.fn = "stop")
}
\arguments{
\item{fdb.fn}{}
}
\value{

}
\description{
Title
}
