% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-check.R
\name{check_no_overlap}
\alias{check_no_overlap}
\title{Check no overlap}
\usage{
check_no_overlap(x, y, fdb.fn = "stop", with.time = FALSE)
}
\arguments{
\item{x, y}{Input vectors whoose content is to be checked.}
}
\value{
Error message if overlap is found. TRUE is no overlap is found.
}
\description{
Check no overlap
}
