% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-check.R
\name{check_none_of}
\alias{check_none_of}
\title{Overwrite check}
\usage{
check_none_of(
  input,
  against,
  ref.input = NULL,
  ref.against,
  overwrite = NULL,
  fdb.fn = "stop",
  with.time = FALSE,
  force = NULL
)
}
\arguments{
\item{input}{An input vector to be checked.}

\item{against}{A vector of valid inputs.}

\item{ref.input}{Character value or NULL. The reference for argument \code{input} input.#'}
}
\description{
Make sure that nothing is overwritten without specifically allowing it.
}
