% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-check.R
\name{check_one_of}
\alias{check_one_of}
\title{Check valid values}
\usage{
check_one_of(input, against, ref.input = NULL)
}
\arguments{
\item{input}{An input vector to be checked.}

\item{against}{A vector of valid inputs.}

\item{ref.input}{Character value or NULL. The reference for argument \code{input} input.#'}
}
\value{
An error message or an invisible TRUE if all values of input are valid.
}
\description{
Check valid values
}
\details{
Error message is build via \code{glue::glue()} building the following
string:

"Value/Values '{invalid_ref}' of {ref.input} is/are invalid. Valid input options are: '{against_ref}'."
}
