% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam-wrapper.R
\name{check_pam_availability}
\alias{check_pam_availability}
\title{Title}
\usage{
check_pam_availability(input, metric.pam, k, fdb.fn = "stop")
}
\arguments{
\item{fdb.fn}{}
}
\value{

}
\description{
Title
}
