% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-check.R
\name{check_vector}
\alias{check_vector}
\title{Compare input to control input}
\usage{
check_vector(
  input,
  against,
  verbose = TRUE,
  ref.input = "input vector",
  ref.against = "against vector"
)
}
\arguments{
\item{input}{A vector of any kind.}

\item{against}{A vector of the same kind as \code{input}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{ref.input}{The reference character value for input.}

\item{ref.against}{The reference character value for against.}
}
\value{
An informative error message about which elements of \code{input} were found in \code{against} or an invisible TRUE.
}
\description{
Compares the values of an input-vector against a control-vector containing
valid values and returns the values of input that were found among the valid ones.
}
