% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-algebra.R
\name{circumference}
\alias{circumference}
\title{Compute circle circumference}
\usage{
circumference(radius, d = Inf)
}
\arguments{
\item{radius}{Numeric vector. The radius/radii.}

\item{d}{Number of digits to which the output values are rounded.}
}
\value{
Numeric vector of the same length.
}
\description{
Computes the circumference of a circle based on it' radius.
}
