% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes-and-spectra.R
\docType{data}
\name{clrp_aaas}
\alias{clrp_aaas}
\title{American Association for the Advancement - color palette}
\format{
An object of class \code{character} of length 10.
}
\usage{
clrp_aaas
}
\description{
American Association for the Advancement - color palette
}
\keyword{datasets}
