% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes-and-spectra.R
\docType{data}
\name{clrp_egr}
\alias{clrp_egr}
\title{Elena Grabis - color palette}
\format{
An object of class \code{character} of length 11.
}
\usage{
clrp_egr
}
\description{
Elena Grabis - color palette
}
\keyword{datasets}
