% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes-and-spectra.R
\docType{data}
\name{clrp_jama}
\alias{clrp_jama}
\title{The Journal of the American Medical Association - color palette}
\format{
An object of class \code{character} of length 10.
}
\usage{
clrp_jama
}
\description{
The Journal of the American Medical Association - color palette
}
\keyword{datasets}
