% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes-and-spectra.R
\docType{data}
\name{clrp_jco}
\alias{clrp_jco}
\title{Journal of Oncology - color panel}
\format{
An object of class \code{character} of length 10.
}
\usage{
clrp_jco
}
\description{
Journal of Oncology - color panel
}
\keyword{datasets}
