% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes-and-spectra.R
\docType{data}
\name{clrp_lo}
\alias{clrp_lo}
\title{Lancet Oncology - color palette}
\format{
An object of class \code{character} of length 10.
}
\usage{
clrp_lo
}
\description{
Lancet Oncology - color palette
}
\keyword{datasets}
