% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes-and-spectra.R
\docType{data}
\name{clrp_milo}
\alias{clrp_milo}
\title{MILO Research Group - color panel}
\format{
An object of class \code{character} of length 20.
}
\usage{
clrp_milo
}
\description{
MILO Research Group - color panel
}
\keyword{datasets}
