% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes-and-spectra.R
\docType{data}
\name{clrp_nejm}
\alias{clrp_nejm}
\title{New England Journal of Medicine- color panel}
\format{
An object of class \code{character} of length 10.
}
\usage{
clrp_nejm
}
\description{
New England Journal of Medicine- color panel
}
\keyword{datasets}
