% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes-and-spectra.R
\docType{data}
\name{clrp_npg}
\alias{clrp_npg}
\title{Nature Publishing Group - color palette}
\format{
An object of class \code{character} of length 10.
}
\usage{
clrp_npg
}
\description{
Nature Publishing Group - color palette
}
\keyword{datasets}
