% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes-and-spectra.R
\docType{data}
\name{clrp_sifre}
\alias{clrp_sifre}
\title{Simon Frerichs - color palette}
\format{
An object of class \code{character} of length 14.
}
\usage{
clrp_sifre
}
\description{
Simon Frerichs - color palette
}
\keyword{datasets}
