% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes-and-spectra.R
\docType{data}
\name{clrp_uc}
\alias{clrp_uc}
\title{University of Chicago - color palette}
\format{
An object of class \code{character} of length 10.
}
\usage{
clrp_uc
}
\description{
University of Chicago - color palette
}
\keyword{datasets}
