% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-Clustering.R
\docType{class}
\name{ClusteringMethod-class}
\alias{ClusteringMethod-class}
\alias{ClusteringMethod}
\title{The \code{ClusteringMethod}-class}
\description{
S4-class that contains results from the clustering
algorithm it stands for.
}
\section{Slots}{

\describe{
\item{\code{key_name}}{character. The name of the variable that is used to identify
each observation uniquely.}

\item{\code{method}}{character. The name of the method. Is additionally encoded
in the name of the S4-class that inherits class \code{ClusteringMethod} using
the syntax \emph{Clustering<method>}.}

\item{\code{results}}{list. A list of results. Varies from cluster method to
cluster method.}
}}

