% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes-and-spectra.R
\docType{data}
\name{color_palettes_md}
\alias{color_palettes_md}
\title{Manually defined palettes}
\format{
An object of class \code{list} of length 14.
}
\usage{
color_palettes_md
}
\description{
Manually defined palettes
}
\keyword{datasets}
