% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes-and-spectra.R
\name{color_vector}
\alias{color_vector}
\title{Get vector of colors}
\usage{
color_vector(clrp, names = NULL, clrp.adjust = NULL)
}
\arguments{
\item{clrp}{Character value. Denotes the colorpanels of interest. Run \code{all_colorpanels()} to
obtain valid inputs.}

\item{names}{Character vector or NULL. Assigns names to the colorvector (in the same order).}

\item{clrp.adjust}{Character vector or NULL. Adjusts the resulting named color vector.
Only named elements are kept. Elements of the color vector with the same names as in
\code{clrp.adjust} are replaced such that the elements of \code{clrp.adjust} are displayed by
the color their name indicates rather than the color the color panel would have assigned
to them.}
}
\value{
Character vector.
}
\description{
Returns a vector of colorcodes of the respective panel
}
