% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-ClusteringKmeans.R
\name{compute_clustering_kmeans}
\alias{compute_clustering_kmeans}
\title{Iterate over kmeans}
\usage{
compute_clustering_kmeans(
  data,
  ks = 2:10,
  methods.kmeans = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{All numeric data.frame or matrix.}

\item{ks}{Numeric vector. All inputs for argument \code{centers} of function
\code{stats::kmeans()}.}

\item{inherit}{argument_dummy params}
}
\value{
A list named according to all specified kmeans methods.
Each of these method-slots contains a list named according to the syntax \emph{k_<k>}
where \emph{<k>} is the respective integer of \code{centers}. The content
is the output of \code{stats::kmeans()}.
}
\description{
Iterate over kmeans
}
