% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-ClusteringPam.R
\name{compute_clustering_pam}
\alias{compute_clustering_pam}
\title{Iterate over pam}
\usage{
compute_clustering_pam(data, ks, methods.pam = NULL, verbose = TRUE, ...)
}
\arguments{
\item{data}{All numeric data.frame or matrix.}

\item{...}{Additional arguments given to \code{cluster::pam()}.}

\item{inherit}{argument_dummy params}
}
\value{
A list named according to all specified pam methods.
Each of these method-slots contains a list named according to the syntax \emph{k_<k>}
where \emph{<k>} is the respective integer of \code{ks}. The content
is the output of \code{cluster::pam()}.
}
\description{
Iterate over pam
}
