% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional-reduction.R
\name{compute_dim_red}
\alias{compute_dim_red}
\title{Title}
\usage{
compute_dim_red(
  data,
  key.name,
  method.dim.red,
  scale = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{scale}{}
}
\value{

}
\description{
Title
}
