% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust-wrapper.R
\name{compute_distance_matrices}
\alias{compute_distance_matrices}
\title{Title}
\usage{
compute_distance_matrices(hcl.obj, methods.dist, p = 2, verbose = TRUE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
