% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust-wrapper.R
\name{compute_distance_matricesx}
\alias{compute_distance_matricesx}
\title{Title}
\usage{
compute_distance_matricesx(
  hcl.obj,
  methods.dist,
  p = 2,
  verbose = TRUE,
  force = FALSE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
