% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust-wrapper.R
\name{compute_hierarchical_cluster}
\alias{compute_hierarchical_cluster}
\title{Title}
\usage{
compute_hierarchical_cluster(
  hcl.obj,
  methods.aggl = NULL,
  methods.dist = NULL,
  verbose = TRUE
)
}
\arguments{
\item{methods.dist}{}
}
\value{

}
\description{
Title
}
