% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Clustering.R
\name{computeClusteringHclust}
\alias{computeClusteringHclust}
\alias{computeClusteringHclust,Clustering-method}
\title{Compute \code{hclust} objects}
\usage{
computeClusteringHclust(object, ...)

\S4method{computeClusteringHclust}{Clustering}(
  object,
  methods_dist = "euclidean",
  methods_aggl = "ward.D",
  verbose = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{methods_dist}{Character value. Specifies the distance methods
of interest. Use \code{validMethodsDist()} to obtain all valid input options.}

\item{methods_aggl}{Character vector. Specifies the agglomerative
methods of interest. Use \code{validMethodsAggl()} to obtain all valid input options.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object.
}
\description{
Computes \code{hclust} objects for all combinations of distance
and agglomration methods. Does not store distance matrices in the object!
}
