% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Clustering.R
\name{computeClusteringPam}
\alias{computeClusteringPam}
\alias{computeClusteringPam,Clustering-method}
\title{Compute cluster with pam}
\usage{
computeClusteringPam(object, ...)

\S4method{computeClusteringPam}{Clustering}(object, ks, methods_pam, verbose = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{ks}{Numeric vector. Denotes all options for k-clusters. Values <1 are discarded.
Is converted into an integer vector.}

\item{methods_pam}{Character value. Specifies the pam-method of interest.
Use \code{validMethodsPam()} to obtain all valid input options.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object.
}
\description{
Computes cluster with method \emph{pam}.
}
