% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-DimRed.R
\name{computePCA}
\alias{computePCA}
\alias{computePCA,DimRed-method}
\title{Compute PCA}
\usage{
computePCA(object, ...)

\S4method{computePCA}{DimRed}(object, n_dims = 2, verbose = TRUE, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{...}{Additional arguments given to \code{irlba::prcomp_irlba()}.}

\item{n_dims}{Numeric value. Integer that indicates the number of
principal components to compute. Must be lower than the number
of numeric variables based on which principal components are
computed.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\description{
Computes principal components and stores the embedding
in form of a data.frame.
}
