% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-DimRed.R
\name{computeTSNE}
\alias{computeTSNE}
\alias{computeTSNE,DimRed-method}
\title{Compute TSNE}
\usage{
computeTSNE(object, ...)

\S4method{computeTSNE}{DimRed}(object, n_dims = 2, verbose = TRUE, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{...}{Additional arguments given to \code{tsne::tsne()}.}

\item{n_dims}{Numeric value. Integer that indicates the number of
dimensions of the resulting embedding. Must be lower than the number
of numeric variables based on which principal components are
computed. Given to argument \code{k} of function \code{tsne::tsne()}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\description{
Computes t stochastic neighbour embedding stores the embedding
in form of a data.frame.
}
