% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-Correlation.R
\name{corr_mtr_to_df}
\alias{corr_mtr_to_df}
\title{Convert correlation matrix to data.frame}
\usage{
corr_mtr_to_df(mtr, type = "complete", diagonal = TRUE, distinct = FALSE)
}
\arguments{
\item{mtr}{A correlation matrix.}

\item{type}{Character value. Denotes how the underlying correlation matrix is
handled. Three options:

\itemize{
\item{\emph{'complete'}:}{ The matrix stays as is.},
\item{\emph{'lower'}:}{ The part below the diagonal is used. The upper part is set to NA.},
\item{\emph{'upper'}:}{The part above the diagonal is used. the lower part is set to NA.}
}}

\item{diagonal}{Logical value.  Indicates if the diagonal values of the
matrices are kept ( = TRUE) or set to NA (= FALSE). Ignored if \code{type}
= \emph{'complete'}.}

\item{distinct}{Logical value. If TRUE only one observation per variable pair remains
in the output data.frame.}
}
\description{
Converts a correlation matrix to a tidy data.frame in
which observations are correlation pairs.
}
