% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-wrapper-progress.R
\name{create_progress_bar}
\alias{create_progress_bar}
\title{Create a progress bar object}
\usage{
create_progress_bar(
  total,
  format = "Progress: [:bar] :percent eta: :eta",
  clear = FALSE,
  width = 100,
  ...
)
}
\arguments{
\item{total}{Numeric value. Total number of ticks.}

\item{format}{Character value. Denotes the format.}

\item{clear}{Logical value.}

\item{width}{Numeric value}
}
\value{
A progress bar object.
}
\description{
Create a progress bar object
}
