% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust-wrapper.R
\name{define_label_params}
\alias{define_label_params}
\title{Title}
\usage{
define_label_params(
  nbLabels,
  labels.angle = 0,
  labels.hjust = 0,
  direction = c("tb", "bt", "lr", "rl"),
  fan = FALSE
)
}
\arguments{
\item{fan}{}
}
\value{

}
\description{
Title
}
