% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-algebra.R
\name{degr2rad}
\alias{degr2rad}
\title{Convert radians to degrees}
\usage{
degr2rad(degr, d = Inf)
}
\arguments{
\item{degr}{Numeric vector in degrees. Values must range from 0 to 360.}

\item{d}{Number of digits to which the output values are rounded.}
}
\value{
Numeric vector of the same length.
}
\description{
Converts radian
}
