% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-OutlierDetection.R
\name{detect_outliers_iqr}
\alias{detect_outliers_iqr}
\title{Detect outliers with IQR}
\usage{
detect_outliers_iqr(df, key.name, var.name)
}
\arguments{
\item{df}{A data.frame.}

\item{key.name}{Character value. The variable that identifies each observation uniquely
with its value.}

\item{var.name}{Character value. The numeric variable according to which
outliers are detected.}
}
\value{
Character vector. The values from variable \code{key.name} that
were identified as outliers. If not outliers exist the function returns
a character vector of length 0.
}
\description{
Uses IQR to detect outliers in a data.frame
}
