% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-Outlier.R
\name{detect_outliers_mahalanobis}
\alias{detect_outliers_mahalanobis}
\title{Detect outliers with Mahalanobis}
\usage{
detect_outliers_mahalanobis(df, key.name, var.names = NULL)
}
\arguments{
\item{df}{A data.frame.}

\item{key.name}{Character value. The variable that identifies each observation uniquely
with its value.}

\item{var.names}{Character vector. The numeric variables according to which
outliers are detected. If NULL, all numeric variable are chosen}
}
\value{
A data.frame with columns \emph{mahal}, \emph{pval} and
\emph{<key.name>}.
}
\description{
Uses mahalanobis distance to detect outliers in a data.frame
}
