% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-OutlierDetection.R
\name{detectOutliersIQR}
\alias{detectOutliersIQR}
\alias{detectOutliersMahalanobis}
\title{Detect outliers}
\usage{
detectOutliersIQR(object, across = NULL, verbose = TRUE)

detectOutliersMahalanobis(object, across = NULL, verbose = TRUE)
}
\arguments{
\item{object}{An object of S4-class \code{OutlierDetection}.}

\item{across}{Character value or NULL. If character,
the data is splittd into the groups the variable \code{across}
contains and outlier detection is conducted separately for
each group.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object.
}
\description{
Runs outlier detection based on the the suffixed method.
}
