% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-DimRed.R
\docType{class}
\name{DimRedMethod-class}
\alias{DimRedMethod-class}
\alias{DimRedMethod}
\title{The \code{DimRedMethod} class}
\arguments{
\item{embedding}{Data.frame. Contains the dimensional reduction embedding.
Variable names correspond to the method suffixed with a number.}

\item{key_name}{Character. Name of the key variable.}

\item{method}{Character. Name of the dimensional reduction method.}
}
\description{
Abstracts conduction of different dimensional reduction
methods or algorithms.
}
