% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-DimRed.R
\docType{class}
\name{DimRedPCA-class}
\alias{DimRedPCA-class}
\alias{DimRedPCA}
\title{The \code{DimRedPCA}}
\arguments{
\item{n_dims}{Numeric value. The number of principal component vectors
returned.}
}
\description{
S4-class that abstracts principal component analysis.
}
\seealso{
\code{DimRedMethod}
}
