% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-DimRed.R
\docType{class}
\name{DimRedTSNE-class}
\alias{DimRedTSNE-class}
\alias{DimRedTSNE}
\title{The \code{DimRedTSNE}}
\description{
S4-class that abstracts t stoachastic neighbour embedding.
}
\seealso{
\code{DimRedMethod}
}
