% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-DimRed.R
\docType{class}
\name{DimRedUMAP-class}
\alias{DimRedUMAP-class}
\alias{DimRedUMAP}
\title{The \code{DimRedUMAP}}
\arguments{
\item{n_dims}{Numeric value. The number of principal component vectors
returned.}
}
\description{
S4-class that abstracts manifold approximation and projection.
}
\seealso{
\code{DimRedMethod}
}
