% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting-aid.R
\name{discard_if}
\alias{discard_if}
\title{Discard values}
\usage{
discard_if(
  input,
  one_of = NULL,
  ref.do = "Ignoring",
  ref.of = "already present",
  ref.input = NULL,
  ref.empty = "Skipping.",
  v.empty = NULL,
  with.time = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{input}{Vector that is to be checked.}

\item{one_of}{Vector that contains the values that must not be inside input for
argument \code{input}.}
}
\description{
Discard values
}
