% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-wrapper.R
\name{discard_numeric_vars}
\alias{discard_numeric_vars}
\title{Discard numeric variables of corr_conv objects}
\usage{
discard_numeric_vars(corr.obj, vars, discard.data = TRUE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Discard numeric variables of corr_conv objects
}
