% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting-aid.R
\name{discard_unnamed}
\alias{discard_unnamed}
\alias{keep_named}
\title{Discard unnamed elements}
\usage{
discard_unnamed(input)

keep_named(input)
}
\arguments{
\item{input}{R object that is to be checked.}
}
\description{
Makes sure that all elements of \code{input} are
named.
}
