% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{error_handler}
\alias{error_handler}
\title{Return function}
\usage{
error_handler(fun)
}
\arguments{
\item{fun}{Character value. One of \emph{'message', 'warning', 'stop'}.}
}
\value{
The respective function
}
\description{
Return function
}
