% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-check.R
\name{extract_feedback}
\alias{extract_feedback}
\title{Title}
\usage{
extract_feedback(fdb.fn)
}
\arguments{
\item{fdb.fn}{Character value. Determines the function to call if a feedback
message needs to be given. One of \emph{'stop', 'warning'} or \emph{'message'}.}
}
\description{
Returns the appropriate string to extract the feedback
from \code{purr::quietly()} results.
}
