% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback-list.R
\docType{data}
\name{feedback_list}
\alias{feedback_list}
\title{A named list of possible feedback messages}
\format{
An object of class \code{list} of length 1.
}
\usage{
feedback_list
}
\description{
A named list of possible feedback messages
}
\keyword{datasets}
