% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math-related-functions.R
\name{fit_model}
\alias{fit_model}
\title{Trigonometric & mathematic curve}
\usage{
fit_model(input, fn, rev = FALSE, normalize = TRUE)
}
\arguments{
\item{input}{Numeric vector.}

\item{fn}{Character value. The \code{confuns::function()} to call.
Currently one of: \emph{'one_peak', 'two_peaks', 'gradient', 'log'.}}

\item{rev}{Logical. If set to TRUE the fitted curve is returned upside-down.}

\item{normalize}{Logical. If set to TRUE numeric values will be scaled down to
values between one and zero.}
}
\value{
A numeric vector of equal length to the input containing the
numeric values of the 'fitted model'.
}
\description{
Fits a curve to the input vector.
}
