% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-wrapper.R
\name{get_corr_results}
\alias{get_corr_results}
\title{Title}
\usage{
get_corr_results(
  corr.obj,
  method.corr = NULL,
  across = NULL,
  across.subset = NULL,
  fdb.fn = "stop"
)
}
\arguments{
\item{fdb.fn}{}
}
\value{

}
\description{
Title
}
