% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust-wrapper.R
\name{get_dendro_data}
\alias{get_dendro_data}
\title{Title}
\usage{
get_dendro_data(
  hcl.obj,
  method.dist = NULL,
  method.aggl = NULL,
  k = NULL,
  h = NULL,
  type = "rectangle"
)
}
\arguments{
\item{h}{}
}
\value{

}
\description{
Title
}
