% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional-reduction.R
\name{get_dim_red_df}
\alias{get_dim_red_df}
\title{Title}
\usage{
get_dim_red_df(dimred.obj, with.data = TRUE, with.meta = TRUE)
}
\arguments{
\item{with.meta}{}
}
\value{

}
\description{
Title
}
