% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_infos.R
\name{get_gene_names}
\alias{get_gene_names}
\alias{get_full_gene_names}
\title{Obtain gene names}
\usage{
get_gene_names(...)

get_full_gene_names(...)
}
\arguments{
\item{...}{Use tidyselect grammer to select specific genes.}
}
\value{
Character vector of HGNC symbols in case of \code{get_gene_names()}
and of full gene names in case of \code{get_full_gene_names()}.
}
\description{
Returns a vector gene names (HGNC symbols). Defaults to all gene names
in the gene info data.frame \code{gene_info_df}.
}
