% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_infos.R
\name{get_gene_synonyms}
\alias{get_gene_synonyms}
\title{Obtain gene synonyms}
\usage{
get_gene_synonyms(gene)
}
\arguments{
\item{gene}{Character value. HGNC symbol of the gene of interest.}
}
\value{
Character vector. NA if there are no synonyms.
}
\description{
Extracts synonyms with which to refer to a gene.
}
