% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust-wrapper.R
\name{get_hclust_df}
\alias{get_hclust_df}
\title{Title}
\usage{
get_hclust_df(
  hcl.obj,
  methods.dist = NULL,
  methods.aggl = NULL,
  k = NULL,
  h = NULL,
  cluster.prefix = "",
  with.data = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{h}{}
}
\value{

}
\description{
Title
}
