% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust-wrapper.R
\name{get_hclust_obj}
\alias{get_hclust_obj}
\title{Title}
\usage{
get_hclust_obj(
  hcl.obj,
  method.dist = NULL,
  method.aggl = NULL,
  fdb.fn = "stop"
)
}
\arguments{
\item{method.aggl}{}
}
\value{

}
\description{
Title
}
