% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans-wrapper.R
\name{get_kmeans_obj}
\alias{get_kmeans_obj}
\title{Title}
\usage{
get_kmeans_obj(
  kmeans.obj,
  method.kmeans = NULL,
  centers = NULL,
  fdb.fn = "stop"
)
}
\arguments{
\item{fdb.fn}{}
}
\value{

}
\description{
Title
}
