% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam-wrapper.R
\name{get_medoids_df}
\alias{get_medoids_df}
\title{Title}
\usage{
get_medoids_df(
  pam.obj,
  k = NULL,
  metric.pam = NULL,
  verbose = TRUE,
  cluster.prefix = "",
  sep = " = ",
  fdb.fn = "message"
)
}
\arguments{
\item{fdb.fn}{}
}
\value{

}
\description{
Title
}
