% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam-wrapper.R
\name{get_pam_df}
\alias{get_pam_df}
\title{Title}
\usage{
get_pam_df(
  pam.obj,
  metric.pam = NULL,
  k = NULL,
  cluster.prefix = "",
  full.name = TRUE,
  sep = "_",
  with.data = FALSE,
  fdb.fn = "message"
)
}
\arguments{
\item{with.data}{}
}
\value{

}
\description{
Title
}
