% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-Analysis.R
\name{getAnalysisAspect}
\alias{getAnalysisAspect}
\alias{getAnalysisAspect,Analysis-method}
\title{Obtain objects of class \code{AnalysisAspect}}
\usage{
getAnalysisAspect(object, ...)

\S4method{getAnalysisAspect}{Analysis}(object, aspect, verbose = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{aspect}{Character value. Name of the analsis aspect. Use \code{validAnalysisAspects()}
to obtain all valid input options.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An object of class \code{AnalysisAspect}.
}
\description{
Extracts objects of class \code{AnalysisAspect} and adds
content of slot @data, @data_scaled and @meta.
}
