% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Clustering.R,
%   R/S4-Analysis.R, R/S4-ClusteringPam.R
\name{getAvgSilWidthsDf}
\alias{getAvgSilWidthsDf}
\alias{getAvgSilWidthsDf,Clustering-method}
\alias{getAvgSilWidthsDf,Analysis-method}
\alias{getAvgSilWidthsDf,ClusteringPam-method}
\title{Obtain average sil-width data}
\usage{
getAvgSilWidthsDf(object, ...)

\S4method{getAvgSilWidthsDf}{Clustering}(object, ks, methods_pam = "euclidean")

\S4method{getAvgSilWidthsDf}{Analysis}(object, ks, methods_pam = "euclidean")

\S4method{getAvgSilWidthsDf}{ClusteringPam}(object, ks, methods_pam = "euclidean")
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{ks}{Numeric vector. Denotes all options for k-clusters. Values <1 are discarded.
Is converted into an integer vector.}
}
\value{
Data.frame.
}
\description{
Extracts a data.frame that contains information
about the average silhouette widths of different clustering
results with mehod \emph{pam}.
}
