% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R
\name{getClustering}
\alias{getClustering}
\title{Obtain \code{Clustering} object}
\usage{
getClustering(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
S4-object of class \code{Clustering}.
}
\description{
Extracts an object of class \code{Clustering} containing
clustering results for all used methods so far.
}
