% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Clustering.R
\name{getClusteringHclust}
\alias{getClusteringHclust}
\alias{getClusteringHclust,Clustering-method}
\title{Obtain object of class \code{ClusteringHclust}}
\usage{
getClusteringHclust(object, ...)

\S4method{getClusteringHclust}{Clustering}(object, ...)

\S4method{getClusteringHclust}{Clustering}(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
An object of class \code{ClusteringHclust}.
}
\description{
Extracts an object of class \code{ClusteringHclust}.
}
\seealso{
getHclust()
}
