% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Clustering.R
\name{getClusteringKmeans}
\alias{getClusteringKmeans}
\alias{getClusteringKmeans,Clustering-method}
\title{Obtain object of class \code{ClusteringKmeans}}
\usage{
getClusteringKmeans(object, ...)

\S4method{getClusteringKmeans}{Clustering}(object, ...)

\S4method{getClusteringKmeans}{Clustering}(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
An object of class \code{ClusteringKmeans}.
}
\description{
Extracts an object of class \code{ClusteringKmeans}.
}
\seealso{
getKmeans()
}
