% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Clustering.R
\name{getClusteringPam}
\alias{getClusteringPam}
\alias{getClusteringPam,Clustering-method}
\title{Obtain object of class \code{ClusteringPam}}
\usage{
getClusteringPam(object, ...)

\S4method{getClusteringPam}{Clustering}(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
An object of class \code{ClusteringPam}.
}
\description{
Extracts an object of class \code{ClusteringPam}.
}
\seealso{
getPam()
}
