% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Clustering.R
\name{getClusterVarsHclust}
\alias{getClusterVarsHclust}
\alias{getClusterVarsHclust,Clustering-method}
\title{Obtain cluster variables (hclust)}
\usage{
getClusterVarsHclust(object, ...)

\S4method{getClusterVarsHclust}{Clustering}(
  object,
  ks = NULL,
  hs = NULL,
  methods_dist = "euclidean",
  methods_aggl = "ward.D",
  prefix = "",
  naming_k = "{method_dist}_{method_aggl}_k{k}",
  naming_h = "{method_dist}_{method_aggl}_h{h}"
)

\S4method{getClusterVarsHclust}{Clustering}(
  object,
  ks = NULL,
  hs = NULL,
  methods_dist = "euclidean",
  methods_aggl = "ward.D",
  prefix = "",
  naming_k = "{method_dist}_{method_aggl}_k{k}",
  naming_h = "{method_dist}_{method_aggl}_h{h}"
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{ks}{Numeric vector. Denotes all options for k-clusters. Values <1 are discarded.
Is converted into an integer vector.}

\item{hs}{Numeric vector. Denotes the heights at which the dendrogram is cut.}

\item{methods_dist}{Character value. Specifies the distance methods
of interest. Use \code{validMethodsDist()} to obtain all valid input options.}

\item{methods_aggl}{Character vector. Specifies the agglomerative
methods of interest. Use \code{validMethodsAggl()} to obtain all valid input options.}
}
\value{
A data.frame.
}
\description{
Extracts a data.frame that contains grouping variables
according to clustering results.
}
