% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Clustering.R,
%   R/S4-Analysis.R
\name{getClusterVarsKmeans}
\alias{getClusterVarsKmeans}
\alias{getClusterVarsKmeans,Clustering-method}
\alias{getClusterVarsKmeans,Analysis-method}
\title{Obtain cluster variables (kmeans)}
\usage{
getClusterVarsKmeans(object, ...)

\S4method{getClusterVarsKmeans}{Clustering}(
  object,
  ks,
  methods_kmeans = "Hartigan-Wong",
  prefix = "",
  naming = "{method_kmeans}_k{k}"
)

\S4method{getClusterVarsKmeans}{Analysis}(
  object,
  ks,
  methods_kmeans = "Hartigan-Wong",
  prefix = "",
  naming = "{method_kmeans}_k{k}"
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{ks}{Numeric vector. Denotes all options for k-clusters. Values <1 are discarded.
Is converted into an integer vector.}
}
\value{
A data.frame.
}
\description{
Extracts a data.frame that contains grouping variables
according to clustering results.
}
