% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Correlation.R,
%   R/S4-Analysis.R
\name{getCorrDf}
\alias{getCorrDf}
\alias{getCorrDf,Correlation-method}
\alias{getCorrDf,Analysis-method}
\alias{getCorrMtr,Analysis-method}
\title{Obtain correlation data.frame}
\usage{
getCorrDf(object, ...)

\S4method{getCorrDf}{Correlation}(
  object,
  method_corr = "pearson",
  across = NULL,
  across_subset = NULL,
  pval_threshold = 0.05,
  type = "complete",
  diagonal = TRUE,
  distinct = FALSE,
  digits = 2,
  verbose = TRUE,
  sep = " & ",
  ...
)

\S4method{getCorrDf}{Analysis}(
  object,
  method_corr = "pearson",
  across = NULL,
  across_subset = NULL,
  pval_threshold = 0.05,
  type = "complete",
  diagonal = TRUE,
  distinct = FALSE,
  digits = 2,
  verbose = TRUE,
  sep = " & ",
  ...
)

\S4method{getCorrMtr}{Analysis}(
  object,
  method_corr = "pearson",
  across = NULL,
  across_subset = NULL,
  type = "complete",
  diagonal = TRUE,
  flatten = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{across}{Character value. Specifies the discrete variable in the data.frame
across which the variables of interest are to be analyzed or displayed.}

\item{pval_threshold}{Numeric value or NULL. If numeric, denotes the maximum p-value to be considered as
significant. Must no be higher than 0.05.}

\item{type}{Character value. Denotes how the underlying correlation matrix is
handled. Three options:

\itemize{
\item{\emph{'complete'}:}{ The matrix stays as is.},
\item{\emph{'lower'}:}{ The part below the diagonal is used. The upper part is set to NA.},
\item{\emph{'upper'}:}{The part above the diagonal is used. the lower part is set to NA.}
}}

\item{diagonal}{Logical value.  Indicates if the diagonal values of the
matrices are kept ( = TRUE) or set to NA (= FALSE). Ignored if \code{type}
= \emph{'complete'}.}

\item{distinct}{Logical value. If TRUE only one observation per variable pair remains
in the output data.frame.}

\item{digits}{Numeric. Given to \code{base::round()} and indicates the number
of digits to which the correlation value is rounded. Defaults to 2.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{sep}{Character value. Denots the string with which the variable pairs
are combined in variable \emph{var_pair} of the output data.frame.}
}
\value{
A data.frame of class \code{corr_df} with the following columns:

\itemize{
\item{\emph{across}:}{ Factor or NULL. If factor, the group names of the grouping variable
denoted in \code{across}.},
\item{\emph{var1}:}{ Factor. First variable of the correlated variable pair.}
\item{\emph{var2}:}{ Factor. Second variable  of the correlated variable pair.},
\item{\emph{var_pair}:}{Factor. Combination of \emph{var1} and \emph{var2}},
\item{\emph{corr}:}{ Numeric. The correlation vaule.},
\item{\emph{pval}:}{ Numeric. The corresponding p-value.},
\item{emph{pval_threshold}:}{ Numeric. The denoted p-value threshold.},
\item{\emph{signif}:}{ Logical. Indicates if the p-value of the correlated pair is below the
denoted threshold.}
\item{\emph{method_corr}:}{ Character. The correlation method.}
}
}
\description{
Extracts correlation results in form of a data.frame
that has been constructed by \emph{melting} a correlation matrix
and its corresponding p-values to a data.frame via \code{reshape2::melt()}.
}
