% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R
\name{getCorrelation}
\alias{getCorrelation}
\title{Obtain \code{Correlation} object}
\usage{
getCorrelation(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
S4-bject of class \code{Correlation}.
}
\description{
Extracts \code{Correlation} object that contains all correlation
results computed so far.
}
