% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Correlation.R
\name{getCorrMtr}
\alias{getCorrMtr}
\alias{getCorrMtr,Correlation-method}
\title{Obtain correlation matrix}
\usage{
getCorrMtr(object, ...)

\S4method{getCorrMtr}{Correlation}(
  object,
  method_corr = "pearson",
  across = NULL,
  across_subset = NULL,
  type = "complete",
  diagonal = TRUE,
  flatten = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{method_corr}{Character value. Specifies the correlation
method of interest. Use \code{validMethodsCorrelation()} to obtain all valid
input options.}

\item{across}{Character value. Specifies the discrete variable in the data.frame
across which the variables of interest are to be analyzed or displayed.}

\item{type}{Character value. Denotes the adjustment type. Either of \emph{'lower',
'upper'} or \emph{'complete'}.}

\item{diagonal}{Logical. Indicates whether the diagonal of the correlation
matrix should be kept or should be converted to NA.}
}
\description{
Extracts correlation results in form of correlation
matrices.
}
\details{
If \code{across} is a character a list of correlation matrices
is returned. Slots are named according to the groups of the denoted
grouping variable. If \code{across_subset} is NULL all groups are considered.
If \code{across_subset} is of length one and \code{flatten} is TRUE the
output list is flattened and a single correlation matrix is returned.
}
