% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Clustering.R
\name{getDendro}
\alias{getDendro}
\alias{getDendro,Clustering-method}
\title{Obtain object of class \code{dendro}}
\usage{
getDendro(object, ...)

\S4method{getDendro}{Clustering}(
  object,
  method_dist = "euclidean",
  method_aggl = "ward.D",
  k = NULL,
  h = NULL,
  type = "rectangle"
)

\S4method{getDendro}{Clustering}(
  object,
  method_dist = "euclidean",
  method_aggl = "ward.D",
  k = NULL,
  h = NULL,
  type = "rectangle"
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{method_dist}{Character value. Specifies the distance method
of interest. Use \code{validMethodsDist()} to obtain all valid input options.}

\item{method_aggl}{Character value. Specifies the agglomerative method
of interest. Use \code{validMethodsAggl()} to obtain all valid input options.}

\item{k}{Numeric value. Denotes the number of clusters. Must be bigger than 1. Must
not be bigger than the number of observations of the data set.}
}
\value{
An object of class \code{dendro}. Contains data for argument
\code{data} of function \code{{ggdendro::ggdendrogram()}}.
}
\description{
Extracts the data from the computed \code{hclust} object
with which a dendrogram can be plotted.
}
