% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Clustering.R
\name{getDendroSegmentDf}
\alias{getDendroSegmentDf}
\alias{getDendroSegmentDf,Clustering-method}
\title{Obtain dendrogram segments}
\usage{
getDendroSegmentDf(object, ...)

\S4method{getDendroSegmentDf}{Clustering}(
  object,
  methods_dist = "euclidean",
  methods_aggl = "ward.D",
  k = NULL,
  h = NULL,
  type = "rectangle"
)

\S4method{getDendroSegmentDf}{Clustering}(
  object,
  methods_dist = "euclidean",
  methods_aggl = "ward.D",
  k = NULL,
  h = NULL,
  type = "rectangle"
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{methods_dist}{Character value. Specifies the distance methods
of interest. Use \code{validMethodsDist()} to obtain all valid input options.}

\item{methods_aggl}{Character vector. Specifies the agglomerative
methods of interest. Use \code{validMethodsAggl()} to obtain all valid input options.}

\item{k}{Numeric value. Denotes the number of clusters. Must be bigger than 1. Must
not be bigger than the number of observations of the data set.}

\item{h}{Numeric value. Denotes the height at which the dendrogram is cut.}
}
\value{
Data.frame.
}
\description{
Extracts a data.frame that contains data to plot
a dendrogram with \code{ggplot2::geom_segment()}.
}
