% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R
\name{getDimRed}
\alias{getDimRed}
\title{Obtain \code{DimRed} object}
\usage{
getDimRed(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
S4-bject of class \code{DimRed}.
}
\description{
Extracts an object of class \code{DimRed} that
contains dimensional reductions computed so far.
}
