% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Clustering.R,
%   R/S4-Analysis.R, R/S4-ClusteringHclust.R
\name{getDistMtr}
\alias{getDistMtr}
\alias{getDistMtr,Clustering-method}
\alias{getDistMtr,Analysis-method}
\alias{getDistMtr,ClusteringHclust-method}
\title{Obtain distance matrix}
\usage{
getDistMtr(object, ...)

\S4method{getDistMtr}{Clustering}(object, method_dist = "euclidean", stop_if_null = FALSE)

\S4method{getDistMtr}{Analysis}(object, method_dist = "euclidean", stop_if_null = FALSE)

\S4method{getDistMtr}{ClusteringHclust}(object, method_dist = "euclidean", stop_if_null = FALSE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{stop_if_null}{Logical value. If TRUE and the function does not find the object to
extract an informative error is raised. Else the empty value - usually NULL - is returned.}
}
\value{
A distance matrix or NULL (if none has been computed yet).
}
\description{
Extracts distance matrix from S4 object.
}
