% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-DimRed.R, R/S4-Analysis.R
\name{getEmbeddingDf}
\alias{getEmbeddingDf}
\alias{getEmbeddingDf,DimRed-method}
\alias{getEmbeddingDf,Analysis-method}
\title{Obtain \code{DimRed} embedding}
\usage{
getEmbeddingDf(object, ...)

\S4method{getEmbeddingDf}{DimRed}(
  object,
  method_dimred = "pca",
  numeric = FALSE,
  numeric_scaled = FALSE,
  grouping = FALSE,
  logical = FALSE,
  complete = FALSE,
  shift = FALSE
)

\S4method{getEmbeddingDf}{Analysis}(
  object,
  method_dimred = "pca",
  numeric = FALSE,
  numeric_scaled = FALSE,
  grouping = FALSE,
  logical = FALSE,
  complete = FALSE,
  sfhit = FALSE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
A data.frame.
}
\description{
Extracts the dimensional reduction embedding in form
of a data.frame. Naming of the variables corresponds
to the chosen method.
}
