% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Clustering.R,
%   R/S4-Analysis.R, R/S4-ClusteringHclust.R
\name{getHclust}
\alias{getHclust}
\alias{getHclust,Clustering-method}
\alias{getHclust,Analysis-method}
\alias{getHclust,ClusteringHclust-method}
\title{Obtain object of class \code{hclust}}
\usage{
getHclust(object, ...)

\S4method{getHclust}{Clustering}(object, method_dist = "euclidean", method_aggl = "ward.D")

\S4method{getHclust}{Analysis}(
  object,
  method_dist = "eucldidean",
  method_aggl = "ward.D",
  stop_if_null = TRUE
)

\S4method{getHclust}{ClusteringHclust}(
  object,
  method_dist = "euclidean",
  method_aggl = "ward.D",
  stop_if_null = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{method_dist}{Character value. Specifies the distance method
of interest. Use \code{validMethodsDist()} to obtain all valid input options.}

\item{method_aggl}{Character value. Specifies the agglomerative method
of interest. Use \code{validMethodsAggl()} to obtain all valid input options.}

\item{stop_if_null}{Logical value. If TRUE and the function does not find the object to
extract an informative error is raised. Else the empty value - usually NULL - is returned.}
}
\value{
An object of class \code{hclust.}
}
\description{
Extracts object of class \code{hclust}
as described in \code{stats::hclust()}.
}
