% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-AnalysisAspect.R,
%   R/S4-Analysis.R
\name{getKeyDf}
\alias{getKeyDf}
\alias{getKeyDf,AnalysisAspect-method}
\alias{getKeyDf,Analysis-method}
\title{Obtain empty data.frame}
\usage{
getKeyDf(object, ...)

\S4method{getKeyDf}{AnalysisAspect}(object, ...)

\S4method{getKeyDf}{Analysis}(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
A data.frame.
}
\description{
Extracts a data.frame that contains only the key variable.
}
