% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Clustering.R,
%   R/S4-Analysis.R, R/S4-ClusteringKmeans.R
\name{getKmeans}
\alias{getKmeans}
\alias{getKmeans,Clustering-method}
\alias{getKmeans,Analysis-method}
\alias{getKmeans,ClusteringKmeans-method}
\title{Obtain object of class \code{kmeans}}
\usage{
getKmeans(object, ...)

\S4method{getKmeans}{Clustering}(object, k, method_kmeans = "Hartigan-Wong")

\S4method{getKmeans}{Analysis}(object, k, method_kmeans = "Hartigan-Wong", stop_if_null = TRUE)

\S4method{getKmeans}{ClusteringKmeans}(object, k, method_kmeans = "Hartigan-Wong", stop_if_null = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{k}{Numeric value. Denotes the number of clusters. Must be bigger than 1. Must
not be bigger than the number of observations of the data set.}

\item{stop_if_null}{Logical value. If TRUE and the function does not find the object to
extract an informative error is raised. Else the empty value - usually NULL - is returned.}
}
\value{
An object of class \code{kmeans}.
}
\description{
Extracts an object of class \code{kmeans}
as described in \code{?stats::kmeans}.
}
\seealso{
getClusteringKmeans()
}
