% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Clustering.R,
%   R/S4-Analysis.R, R/S4-ClusteringPam.R
\name{getMedoidsDf}
\alias{getMedoidsDf}
\alias{getMedoidsDf,Clustering-method}
\alias{getMedoidsDf,Analysis-method}
\alias{getMedoidsDf,ClusteringPam-method}
\title{Obtain medoids from PAM clustering}
\usage{
getMedoidsDf(object, ...)

\S4method{getMedoidsDf}{Clustering}(
  object,
  ks,
  methods_pam = "euclidean",
  prefix = "",
  format = "wide"
)

\S4method{getMedoidsDf}{Analysis}(
  object,
  ks,
  methods_pam = "euclidean",
  prefix = "",
  format = "wide"
)

\S4method{getMedoidsDf}{ClusteringPam}(
  object,
  ks,
  methods_pam = "euclidean",
  prefix = "",
  format = "wide"
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{ks}{Numeric vector. Denotes all options for k-clusters. Values <1 are discarded.
Is converted into an integer vector.}
}
\description{
Extracts a data.frame containing medoids from PAM clustering
results
}
