% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{getMetaNames}
\alias{getMetaNames}
\alias{getMetaNames,Analysis-method}
\title{Obtain meta names}
\usage{
getMetaNames(object)

\S4method{getMetaNames}{Analysis}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
Character vector.
}
\description{
Extracts names of meta variables.
}
