% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-AnalysisAspect.R
\name{getMtr}
\alias{getMtr}
\alias{getMtr,AnalysisAspect-method}
\title{Obtain data matrix}
\usage{
getMtr(object, ...)

\S4method{getMtr}{AnalysisAspect}(object, scale = FALSE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\description{
Extracts the numeric data of the object in form
of a matrix. The key variable is used as rownames.
}
