% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-outlierDetection.R
\name{getOutlierResults}
\alias{getOutlierResults}
\alias{getOutlierResults,OutlierDetection-method}
\title{Obtain outlier results}
\usage{
getOutlierResults(object, method = "IQR", across = NULL, verbose = TRUE, ...)

\S4method{getOutlierResults}{OutlierDetection}(object, method = "IQR", across = NULL, verbose = TRUE)
}
\arguments{
\item{method}{Character value. The outlier detection method. Valid
input options are \emph{'IQR'} and \emph{'Mahalanobis'}.}

\item{across}{Character value or NULL. If character, denotes
the grouping variable of interest.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
If \code{across} is NULL the return value is a data.frame if
\code{method} = \emph{'Mahalanobis'} or a list if \code{method} = \emph{'IQR'}.
In this list each slot is a vector of IDs that were identified
as outliers within the variable denoted by the name of the list's slot.

If \code{across} is a character value the returned value is a list
named according to the groups of the grouping variable \code{across} each containing
output as described above.
}
\description{
Extracts outlier results from the object.
}
