% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-outlierDetection.R
\name{getOutliers}
\alias{getOutliers}
\alias{getOutliers,OutlierDetection-method}
\title{Obtain outlier results}
\usage{
getOutliers(object, method = "IQR", across = NULL, verbose = TRUE, ...)

\S4method{getOutliers}{OutlierDetection}(object, method = "IQR", across = NULL, verbose = TRUE)
}
\arguments{
\item{method}{Character value. The outlier detection method. Valid
input options are \emph{'IQR'} and \emph{'Mahalanobis'}.}

\item{across}{Character value or NULL. If character, denotes
the grouping variable of interest.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
If \code{across} is a character value the returned value is a list
named according to the groups of the grouping variable \code{across}. Else
its a data.frame if \code{method} = \emph{'Mahalanobis'} or a list if
\code{method} = \emph{'IQR'}.
}
\description{
Extracts outlier results from the object.
}
