% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Clustering.R,
%   R/S4-Analysis.R, R/S4-ClusteringPam.R
\name{getPam}
\alias{getPam}
\alias{getPam,Clustering-method}
\alias{getPam,Analysis-method}
\alias{getPam,ClusteringPam-method}
\title{Obtain object of class \code{pam}}
\usage{
getPam(object, ...)

\S4method{getPam}{Clustering}(object, k, method_pam = "euclidean", stop_if_null = TRUE)

\S4method{getPam}{Analysis}(object, k, method_pam = "euclidean", stop_if_null = TRUE)

\S4method{getPam}{ClusteringPam}(object, k, method_pam = "euclidean", stop_if_null = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{k}{Numeric value. Denotes the number of clusters. Must be bigger than 1. Must
not be bigger than the number of observations of the data set.}

\item{stop_if_null}{Logical value. If TRUE and the function does not find the object to
extract an informative error is raised. Else the empty value - usually NULL - is returned.}
}
\value{
An object of class \code{pam}.
}
\description{
Extracts an object of class \code{pam}
as described in \code{?cluster::pam}.
}
\seealso{
getClusteringPam()
}
