% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Correlation.R,
%   R/S4-Analysis.R
\name{getRcorr}
\alias{getRcorr}
\alias{getRcorr,Correlation-method}
\alias{getRcorr,CorrelationMethod-method}
\alias{getRcorr,Analysis-method}
\title{Obtain object of class \code{rcorr}}
\usage{
getRcorr(object, ...)

\S4method{getRcorr}{Correlation}(
  object,
  method_corr = "pearson",
  across = NULL,
  across_subset = NULL,
  as_list = FALSE,
  flatten = TRUE
)

\S4method{getRcorr}{CorrelationMethod}(
  object,
  method_corr = "pearson",
  across = NULL,
  across_subset = NULL,
  as_list = FALSE,
  stop_if_null = TRUE
)

\S4method{getRcorr}{Analysis}(
  object,
  method_corr = "pearson",
  across = NULL,
  across_subset = NULL,
  as_list = FALSE,
  stop_if_null = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{across}{Character value. Specifies the discrete variable in the data.frame
across which the variables of interest are to be analyzed or displayed.}

\item{as_list}{Logical value. If TRUE, \code{rcorr} objects are returned
as simple lists.}

\item{stop_if_null}{Logical value. If TRUE and the function does not find the object to
extract an informative error is raised. Else the empty value - usually NULL - is returned.}
}
\description{
Extracts correlation results in form of class \code{rcorr}.
}
\details{
If \code{across} is NULL one object of class \code{rcorr} is
returned - the one for the complete data set. If \code{across} is a character the list of \code{rcorr} is
extracted and subsetted by the groups of interest denoted in \code{across_subset}.
If \code{across_subset} is NULL the complete list is returned. If \code{across_subset}
is of length one and \code{flatten} is TRUE the output list is flattened and
a single \code{rcorr} object is returned.
}
