% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-AnalysisAspect.R
\name{getVariableNames}
\alias{getVariableNames}
\alias{getVariableNames,AnalysisAspect-method}
\title{Obtain variable names}
\usage{
getVariableNames(object, ...)

\S4method{getVariableNames}{AnalysisAspect}(
  object,
  types = c("key", "numeric", "grouping", "logical", "meta"),
  unname = FALSE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
Character vector.
}
\description{
Extracts the variable names of the respective slots in
form of a character vector.
}
