% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{ggplot2_dummy}
\alias{ggplot2_dummy}
\title{ggplot2_dummy}
\usage{
ggplot2_dummy()
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
ggplot2_dummy
}
