% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-wrapper-ggpubr.R
\name{ggpubr_comparison_list}
\alias{ggpubr_comparison_list}
\alias{ggpubr_y_labels}
\title{Prepare input for \code{ggpubr::stat_compare_means()}}
\usage{
ggpubr_comparison_list(ref.group, groups)

ggpubr_y_labels(lst, max.value)
}
\description{
Prepare input for \code{ggpubr::stat_compare_means()}
}
