% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes-and-spectra.R
\docType{data}
\name{ggsci_list}
\alias{ggsci_list}
\title{@rdname milo_clrp
 @export}
\format{
An object of class \code{list} of length 7.
}
\usage{
ggsci_list
}
\description{
@rdname milo_clrp
 @export
}
\keyword{datasets}
