% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-check.R
\name{give_feedback}
\alias{give_feedback}
\title{Print feedback in console}
\usage{
give_feedback(
  fdb.fn = "message",
  msg = NULL,
  in.shiny = FALSE,
  with.time = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{fdb.fn}{Character value. Determines the function to call if a feedback
message needs to be given. One of \emph{'stop', 'warning'} or \emph{'message'}.}

\item{msg}{Character value or glue. The message to be printed in the console.}

\item{in.shiny}{Allows to use the function to stop a function without crashing
a shiny session.}

\item{with.time}{Logical value. Indicates whether the current time is to be
added to the feedback message.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\description{
Helper that gives feedback with a function of choice.
}
