% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{glue_list_report}
\alias{glue_list_report}
\title{Glue a human readable list report}
\usage{
glue_list_report(lst, separator = " = ")
}
\arguments{
\item{lst}{A named list of values.}

\item{separator}{Character value.}
}
\value{
Glue object.
}
\description{
Glue a human readable list report
}
