% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{hclust_dummy}
\alias{hclust_dummy}
\title{hclust_dummy}
\usage{
hclust_dummy(type)
}
\arguments{
\item{type}{Character value. Indicates the shape of the dendrogram.
Input \emph{'rectangle'} will draw rectangular lines, while \emph{'triangle'} will draw triangular lines.}
}
\description{
hclust_dummy
}
