% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-wrapper-shinyjqui.R
\name{hlpr_order_input}
\alias{hlpr_order_input}
\title{Work around}
\usage{
hlpr_order_input(order_input)
}
\description{
Awkward solution to the problem that
input$change_order_order (of shinyjqui::orderInput()) somehow changes it's class
}
