% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience-functions.R
\name{hm}
\alias{hm}
\title{Print top n matrix rows and columns}
\usage{
hm(mtr, n = 6)
}
\arguments{
\item{mtr}{A matrix.}

\item{n}{Numeric value.}
}
\value{
Head of matrix.
}
\description{
Print top n matrix rows and columns
}
