% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-wrapper.R
\name{initiate_corr_object}
\alias{initiate_corr_object}
\title{Title}
\usage{
initiate_corr_object(
  corr.data,
  default.method = c("pearson", "spearman"),
  default.dir = "conv-corr-obj.RDS"
)
}
\arguments{
\item{default.dir}{}
}
\value{

}
\description{
Title
}
