% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans-wrapper.R
\name{initiate_kmeans_object}
\alias{initiate_kmeans_object}
\title{Title}
\usage{
initiate_kmeans_object(
  kmeans.data,
  key.name,
  default.method.kmeans = "Hartigan-Wong",
  default.centers = 2,
  default.dir = "conv-kmeans-obj.RDS"
)
}
\arguments{
\item{default.centers}{}
}
\value{

}
\description{
Title
}
