% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam-wrapper.R
\name{initiate_pam_object}
\alias{initiate_pam_object}
\title{Title}
\usage{
initiate_pam_object(
  pam.data,
  key.name,
  default.as.dist = FALSE,
  default.metric.pam = "euclidean",
  default.k = 2,
  default.dir = "conv-pam-obj.RDS",
  verbose = TRUE
)
}
\arguments{
\item{default.dir}{}
}
\value{

}
\description{
Title
}
