% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-Analysis.R
\name{initiateAnalysis}
\alias{initiateAnalysis}
\title{Initiate analysis}
\usage{
initiateAnalysis(
  data,
  key_name,
  key_prefix = NULL,
  meta_names = character(0),
  cluster_with = "data",
  lgl_to_group = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{Data.frame containing the data to be analyzed.}

\item{key_name}{The key variable. If NULL, the key variable is created
either by using the rownames and - if rownames are invalid - by combining
input for argument \code{key_prefix} with the rownumbers.}

\item{key_prefix}{Character value. The prefix for the artificial
key variable.}

\item{meta_names}{Names of the data.frame of \code{data} that are supposed
to be treated as meta data. Meta data is not integrated in any form
of analysis.}

\item{verbose}{}

\item{analysis_aspect}{The actual analysis aspect. Use \code{validAnalysisAspects()}
to obtain all valid input options.}
}
\value{
An object of class specified in \code{analysis_aspect}.
}
\description{
Sets up an object of class \code{AnalysisAspect}.
}
