% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-AnalysisAspect.R, R/S4-DimRed.R,
%   R/S4-Clustering.R, R/S4-Correlation.R
\name{initiateAnalysisAspect}
\alias{initiateAnalysisAspect}
\alias{initiateDimRed}
\alias{initiateClustering}
\alias{initiateCorrelation}
\title{Initiate analysis}
\usage{
initiateAnalysisAspect(
  data,
  key_name,
  key_prefix = NULL,
  meta_names = character(0),
  lgl_to_group = TRUE,
  analysis_aspect = "AnalysisAspect",
  verbose = TRUE
)

initiateDimRed(
  data,
  key_name,
  key_prefix = NULL,
  lgl_to_group = TRUE,
  meta_names = character(0),
  verbose = TRUE
)

initiateClustering(
  data,
  key_name,
  key_prefix = NULL,
  lgl_to_group = TRUE,
  meta_names = character(0),
  verbose = TRUE
)

initiateCorrelation(
  data,
  key_name,
  key_prefix = "ID",
  lgl_to_group = TRUE,
  meta_names = character(0),
  verbose = TRUE
)
}
\arguments{
\item{data}{Data.frame containing the data to be analyzed.}

\item{key_name}{The key variable. If NULL, the key variable is created
either by using the rownames and - if rownames are invalid - by combining
input for argument \code{key_prefix} with the rownumbers.}

\item{key_prefix}{Character value. The prefix for the artificial
key variable.}

\item{meta_names}{Names of the data.frame of \code{data} that are supposed
to be treated as meta data. Meta data is not integrated in any form
of analysis.}

\item{analysis_aspect}{The actual analysis aspect. Use \code{validAnalysisAspects()}
to obtain all valid input options.}

\item{verbose}{}
}
\value{
An object of class specified in \code{analysis_aspect}.
}
\description{
Sets up an object of class \code{AnalysisAspect}.
}
