% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-outlierDetection.R
\name{initiateOutlierDetection}
\alias{initiateOutlierDetection}
\title{Set up \code{OutlierDetection} object}
\usage{
initiateOutlierDetection(data, key_name, key_prefix = "ID", verbose = TRUE)
}
\arguments{
\item{data}{A data.frame. (Rownames are dropped.)}

\item{key_name}{Character value or NULL. Denotes the variable that is used
to identify each observation uniquely.}

\item{key_prefix}{Character value. The string with which
the IDs in the newly constructed key variable are prefixed if \code{key_name} is
NULL.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An object of S4-class \code{OutlierDetection}.
}
\description{
Sets up an object of class \code{OutlierDetection}.
}
\details{
If \code{key_name} is NULL the key variable is constructed
as a combination of the input data.frames rownames and the string
provided with \code{key_prefix}. If the the datas rownames are NULL
or contain \emph{""} the rownumbers are used instead. The newly constructed
key variable is named \emph{data_ids}.
}
