% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-check.R
\name{is_any_of}
\alias{is_any_of}
\title{Any input check}
\usage{
is_any_of(input, valid.classes)
}
\arguments{
\item{input}{Object to be checked.}

\item{valid.classes}{Character vector to specify the classes the input
might have in order not to be discarded.}
}
\value{
TRUE if \code{input} is of at least one of
the specified classes in \code{valid.classes}. FALSE
if not.
}
\description{
Any input check
}
