% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting-aid.R
\name{is_named}
\alias{is_named}
\title{Logical naming test}
\usage{
is_named(input)
}
\arguments{
\item{input}{R object that is to be checked.}
}
\description{
Returns TRUE if \code{input}'s names evaluate to truthy.
}
