% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join-lists.R
\name{join_lists}
\alias{join_lists}
\title{Join two lists by their names}
\usage{
join_lists(lst.1 = list(), lst.2 = list(), drop.unnamed = TRUE)
}
\arguments{
\item{lst.1}{A named list.}

\item{lst.2}{A named list.}

\item{drop.unnamed}{Logical. If set to TRUE unnamed elements of the return list
will be dropped.}
}
\value{
A merged list.
}
\description{
Overwrites elements of \code{list.1} with those from \code{list.2}
that share the same name. Then merges all residual \strong{named} elements of
\code{list.2} into \code{list.1}. Unnamed elements of \code{list.2} will be dropped.
}
