% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam-clustering.R
\name{k_dummy}
\alias{k_dummy}
\title{k_dummy}
\usage{
k_dummy()
}
\arguments{
\item{k}{Numeric value. Positive integer specifiying the number of clusters.
Must be lower than the number of observations in the supplied data.frame.}
}
\description{
k_dummy
}
