% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-themes-and-small-add-ons.R
\name{legend_none}
\alias{legend_none}
\alias{legend_right}
\alias{legend_left}
\alias{legend_bottom}
\title{Legend location}
\usage{
legend_none(...)

legend_right(...)

legend_left(...)

legend_bottom(...)
}
\description{
Legend location
}
