% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_and_load.R
\name{load_files}
\alias{load_files}
\title{Load in all files at once}
\usage{
load_files(
  folder,
  file.types = c(".csv", ".xls", ".rds", ".xlsx", ".RData"),
  file.subset = character(),
  file.ignore = character(),
  read.funs = list(),
  recursive = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{folder}{The absolute or relative to the folder of interest.}

\item{file.types}{The file types you want to load in specified as a character
vector. Elements must start with '.' ('.csv', NOT 'csv')!}

\item{file.subset}{A regular expression that match the files you want to load
in specified as a character vector.}

\item{file.ignore}{A regular expression that match the files you want to be
ignored specified as a character vector.}

\item{read.funs}{A named list providing the functions that are supposed to be
used to load the respective file type. Names need to match the respective elements
of argument \code{file.types} exactly.}

\item{recursive}{Logical. If set to TRUE the listing recurs into subsequent
folders.}

\item{verbose}{Logical. If set to TRUE informative messages will be printed.

(Warnings will always be printed.)}
}
\value{
Loaded files  in the global environment as a side effect and returns
an invisible logical value: TRUE.
}
\description{
Loads or reads in all files of the respective folder according to the
parameters input.
}
