% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_and_load.R
\name{load_functions}
\alias{load_functions}
\title{Load in all functions}
\usage{
load_functions(folder = "R", recursive = FALSE, verbose = TRUE)
}
\arguments{
\item{folder}{The absolute or relative to the folder of interest.}

\item{recursive}{Logical. If set to TRUE the file listing recurs into subsequent
folders (starting with the folder specified in \code{folder}).}

\item{verbose}{Logical. If set to TRUE informative messages will be printed.

(Warnings will always be printed.)}
}
\value{
Sourced in functions as a side effect and an invisible TRUE.
}
\description{
Sources all R-scripts from specified folder via \code{base::source()} that end with
\emph{_function.R} or \emph{_functions.R}.
}
