% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source-and-load.R
\name{load_scripts}
\alias{load_scripts}
\title{Load in all R scripts}
\usage{
load_scripts(
  folder = "R",
  script.subset = NULL,
  recursive = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{folder}{The absolute or relative to the folder of interest.}

\item{script.subset}{A regular expression that match the file names you want to load
in specified as a character vector.}

\item{recursive}{Logical. If set to TRUE the file listing recurs into subsequent
folders (starting with the folder specified in \code{folder}).}

\item{verbose}{Logical. If set to TRUE informative messages will be printed.

(Warnings will always be printed.)}
}
\value{
Sourced in functions as a side effect and an invisible TRUE.
}
\description{
Sources all R-scripts from specified folder via \code{base::source()}.
}
