% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-manipulation.R
\name{make_capital_letters}
\alias{make_capital_letters}
\title{Capitalizes all words of a string}
\usage{
make_capital_letters(
  string,
  split.by = "_",
  collapse.with = " ",
  capital.letters = TRUE
)
}
\arguments{
\item{string}{Character value.}

\item{split.by}{Character value. Specfies the regex by which the
words of the specified string are split.}

\item{collapse.with}{Character value or \code{NULL}. If character, specifies the string with
which all capitalized words of the input string are reunited to the
output string. (Default is an empty space.)}

\item{capital.letters}{Logical. If set to FALSE capitalization is skipped.}
}
\value{
A more aesthetically pleasing version of the input.
}
\description{
Splits the string in single words by the
regex specified in \code{split.by} and capitalizes every
resulting word which are eventually collapsed with the
string specified in \code{collapse.with}.
}
