% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-manipulation.R
\name{make_pretty_df}
\alias{make_pretty_df}
\title{Make a data.frame more aesthetically pleasing}
\usage{
make_pretty_df(
  df,
  column.names = TRUE,
  group.names = TRUE,
  make.pretty = TRUE,
  ...
)
}
\arguments{
\item{df}{A data.frame.}

\item{column.names}{Logical. If FALSE, column names are skipped.}

\item{group.names}{Logical. If FALSE, discrete variables are skipped.}

\item{make.pretty}{Logical. If FALSE, all functions are skipped.}

\item{...}{Additional arguments given to \code{make_pretty_names}.}
}
\value{
A more aesthetically pleasing version of the input.
}
\description{
Make a data.frame more aesthetically pleasing
}
