% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-manipulation.R
\name{make_pretty_name}
\alias{make_pretty_name}
\alias{make_pretty_names}
\alias{make_pretty_levels}
\title{Make string aesthetically pleasing}
\usage{
make_pretty_name(string, make.pretty = TRUE, ...)

make_pretty_names(strings, ...)

make_pretty_levels(fct, ...)
}
\arguments{
\item{string}{Character value. Returns input as is if it is not a character value.}

\item{make.pretty}{Logical. If set to FALSE the input string is returned.}

\item{...}{Additional argument given to \code{make_capital_letters}.}

\item{strings}{Character vector.}

\item{fct}{Factor.}
}
\value{
A more aesthetically pleasing version of the input.
}
\description{
Takes input and applies the following functions to it

\enumerate{
\item{\code{make_capital_letters()}}
}
}
