% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-manipulation.R
\name{make_pretty_names}
\alias{make_pretty_names}
\title{Make character values aesthethically pleasing}
\usage{
make_pretty_names(string, make.pretty = TRUE)
}
\arguments{
\item{string}{Character value to process.}

\item{pretty.names}{Logical value.}
}
\value{

}
\description{
Make character values aesthethically pleasing
}
