% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-Correlation.R
\name{melt_rcorr}
\alias{melt_rcorr}
\title{Melt \code{rcorr}}
\usage{
melt_rcorr(rcorr_obj, type = "complete", diagonal = TRUE, distinct = FALSE)
}
\arguments{
\item{rcorr_obj}{An object of class \code{rcorr}.}

\item{type}{Character value. Denotes the adjustment type. Either of \emph{'lower',
'upper'} or \emph{'complete'}.}

\item{diagonal}{Logical. Indicates whether the diagonal of the correlation
matrix should be kept or should be converted to NA.}
}
\value{
A data.frame with the following columns:

\itemize{
 \item{\emph{var1}:}{ Factor. First one of the variable pair.}
 \item{\emph{var2}:}{ Factor. Second one of the correlated variable pair.},
 \item{\emph{corr}:}{ Numeric. The correlation vaule.},
 \item{\emph{pval}:}{ Numeric. The corresponding p-value.},
 }
}
\description{
Melts object of class \code{rcorr} to a data.frame.
}
