% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes-and-spectra.R
\docType{data}
\name{milo_clrp}
\alias{milo_clrp}
\title{Color panels}
\format{
An object of class \code{character} of length 25.
}
\usage{
milo_clrp
}
\description{
Color panels
}
\keyword{datasets}
