% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-leverage.R
\name{mrename}
\alias{mrename}
\alias{mrename_with}
\title{Rename matrix in tidyverse style}
\usage{
mrename(mtr, dims = c(1, 2), ...)

mrename_with(mtr, dims = c(1, 2), .fn, ...)
}
\arguments{
\item{mtr}{}

\item{dims}{}

\item{...}{}
}
\value{

}
\description{
Rename matrix in tidyverse style
}
