% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam-clustering.R
\name{n_dummy}
\alias{n_dummy}
\title{n_dummy}
\usage{
n_dummy(n)
}
\arguments{
\item{n}{The max value for \code{k} of \code{cluster::pam()}. The function
computes the silhouette widhts for all pam-results from 2 up to \code{n}.}
}
\description{
n_dummy
}
