% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math-related-functions.R
\name{normalize}
\alias{normalize}
\title{Normalize numeric values}
\usage{
normalize(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric with the same length as \code{x}.
}
\description{
Scales the values of the provided
vector to values between 0 and 1.
}
\examples{
purrr::map_df(.x = mtcars[,sapply(mtcars, is.numeric)],
              .f = normalize)
}
