% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{normalize_dummy}
\alias{normalize_dummy}
\title{normalize_dummy}
\usage{
normalize_dummy()
}
\arguments{
\item{normalize}{Logical. If set to TRUE numeric values will be rescaled to
values between one and zero.}
}
\description{
normalize_dummy
}
