% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math-related-functions.R
\name{normalize_zscore}
\alias{normalize_zscore}
\title{Zscore computation}
\usage{
normalize_zscore(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric with the same length as \code{x}.
}
\description{
Computes zscores of the provided variable.
}
