% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math-related-functions.R
\name{one_peak}
\alias{one_peak}
\alias{two_peaks}
\alias{linear}
\alias{sinus}
\alias{gradient}
\alias{log}
\title{Fit a curve}
\usage{
one_peak(input)

two_peaks(input)

linear(input)

sinus(input)

gradient(input)

log(input)
}
\arguments{
\item{input}{Numeric vector.}
}
\description{
Functions that return the numeric values of the respective
mathematic function they are constructed of. Length and range of the
return vector is equal to those of the input vector.
}
