% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-outlierDetection.R
\docType{class}
\name{OutlierDetection-class}
\alias{OutlierDetection-class}
\alias{OutlierDetection}
\title{The \code{OutlierDetection}-class}
\description{
S4-class that contains data and the corresponding outlier detection
results of several methods.
}
\section{Slots}{

\describe{
\item{\code{data}}{data.frame. The data based on which outlier detection is conducted.}

\item{\code{key_name}}{character. The name of the variable that is used to identify
each observation uniquely.}

\item{\code{methods}}{list. A list of objects of S4-class \code{OutlierDetectionMethods}.}

\item{\code{variables_grouping}}{character. The names of all grouping variables
of the input data - variables of class character or factor. (Does not include
variable of slot @key_name)}

\item{\code{variables_numeric}}{character. The names of all numeric variables
based on which outlier detection is conducted.}
}}

