% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-Outlier.R
\docType{class}
\name{OutlierDetectionIQR-class}
\alias{OutlierDetectionIQR-class}
\alias{OutlierDetectionIQR}
\title{The \code{OutlierDetectionIQR}-class}
\description{
S4-class that contains results from outlier detection
algorithm \emph{Interquartile Range}.
}
\section{Slots}{

\describe{
\item{\code{results}}{list. Named list of IDs that were identified as outliers.
IDs are extracted from the key variable.}

\item{\code{results_across}}{list (nested list). Named list of lists as described
in slot @results. Named according to the grouping variables across which
outlier detection has been conducted.}
}}

