% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-OutlierDetection.R
\docType{class}
\name{OutlierDetectionMahalanobis-class}
\alias{OutlierDetectionMahalanobis-class}
\alias{OutlierDetectionMahalanobis}
\title{The \code{OutlierDetectionMahalanobis}-class}
\description{
S4-class that contains results from outlier detection
algorithm \emph{Mahalanobis}.
}
\section{Slots}{

\describe{
\item{\code{results}}{data.frame. A data.frame with columns \emph{mahal}, \emph{pval} and
\emph{<key_name>}.}

\item{\code{results_across}}{list. Named list of data.frames as described
in slot @results. Named according to the grouping variables across which
outlier detection has been conducted.}
}}

