% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-Outlier.R
\docType{class}
\name{OutlierDetectionMethod-class}
\alias{OutlierDetectionMethod-class}
\alias{OutlierDetectionMethod}
\title{The \code{OutlierDetectionMethod}-class}
\description{
S4-class that contains results from the outlier detection
algorithm it stands for.
}
\section{Slots}{

\describe{
\item{\code{key_name}}{character. The name of the variable that is used to identify
each observation uniquely.}

\item{\code{method}}{character. The name of the method. Is additionally encoded
as a suffix in the S4-class that inherits class \code{OutlierDetectionMethod}}
}}

