% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam-clustering.R
\name{pam_add_cluster}
\alias{pam_add_cluster}
\title{PAM - Clustering}
\usage{
pam_add_cluster(
  df,
  k,
  var.name = "pam_cluster",
  var.type = "factor",
  assign = FALSE,
  assign.name = "pam_obj",
  ...
)
}
\arguments{
\item{df}{A data.frame.}

\item{k}{Numeric value. Positive integer specifying the number of clusters.
Must be lower than the number of observations in the supplied data.frame.}

\item{var.name}{Character value. Specifying the string of the variable
containing the clusters.}

\item{var.type}{Characte value. Specifying the class of the variable
containing the clusters. Must be one of \emph{'numeric', 'factor', 'character'}.}

\item{assign}{Logical.}

\item{...}{Additional arguments given to \code{clusters::pam()}.}
}
\value{
The input data.frame specified in \code{df} with an additional
variable containing the clusters information.
}
\description{
Adds a pam-cluster variable to the supplied
data.frame.
}
\details{
Performs clustering by including all numeric variables.
}
