% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam-clustering.R
\name{pam_plot_elbow}
\alias{pam_plot_elbow}
\title{PAM - Plot Silhouette Widths}
\usage{
pam_plot_elbow(df, n = 10, ...)
}
\arguments{
\item{df}{A data.frame.}

\item{n}{The max value for \code{k} of \code{cluster::pam()}. The function
computes the silhouette widhts for all pam-results from 2 up to \code{n}.}

\item{...}{Additional arguments given to \code{cluster::pam()}.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Takes a data.frame and plots the pam-silhouette-widths
for several k-values.
}
