% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam-clustering.R
\name{pam_plot_silinfo}
\alias{pam_plot_silinfo}
\title{PAM - Silhouette Width Info}
\usage{
pam_plot_silinfo(pam.obj)
}
\arguments{
\item{pam.obj}{A valid pam-object.}

\item{legend.position}{Character. Given to \code{ggplot2::theme()}.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Plots the silhouette information of
the provided pam-object.
}
