% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam-wrapper.R
\name{plot_avg_silhouette_widths}
\alias{plot_avg_silhouette_widths}
\title{Title}
\usage{
plot_avg_silhouette_widths(
  pam.obj,
  k,
  metric.pam = NULL,
  clr = "steelblue",
  display.cols = TRUE,
  display.line = TRUE,
  display.points = TRUE
)
}
\arguments{
\item{display.points}{}
}
\value{

}
\description{
Title
}
