% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-statistics-interactive.R
\name{plot_categorical_statistics_shiny}
\alias{plot_categorical_statistics_shiny}
\title{Title}
\usage{
plot_categorical_statistics_shiny(
  df,
  variables,
  across,
  across.subset = NULL,
  clrp = "milo",
  position = "fill",
  ...
)
}
\value{

}
\description{
Title
}
