% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-wrapper.R
\name{plot_correlation_sd}
\alias{plot_correlation_sd}
\alias{plot_correlation_variance}
\title{Title}
\usage{
plot_correlation_sd(
  corr.obj,
  method.corr = NULL,
  across = NULL,
  aes.fill = "sd",
  signif.level = NULL
)

plot_correlation_variance(
  corr.obj,
  method.corr = NULL,
  across = NULL,
  aes.fill = "sd",
  signif.level = NULL
)
}
\arguments{
\item{aes.fill}{}
}
\value{

}
\description{
Title
}
