% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-wrapper.R
\name{plot_corrplot}
\alias{plot_corrplot}
\title{Title}
\usage{
plot_corrplot(
  corr.input,
  method.corr = NULL,
  plot.type = "lower",
  display.diagonal = TRUE,
  p.mtr = NULL,
  signif.level = NULL,
  clr.low = "darkred",
  clr.high = "steelblue",
  shape = "tile",
  display.with.size = TRUE,
  size.max = 15,
  display.values = TRUE,
  values.alpha = 0.9,
  values.clr = "black",
  values.digits = 2,
  values.size = 4,
  draw.grid = TRUE,
  grid.clr = "black",
  grid.size = 0.5
)
}
\arguments{
\item{grid.size}{}
}
\value{

}
\description{
Title
}
