% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust-wrapper.R
\name{plot_dendrogram}
\alias{plot_dendrogram}
\alias{plot_dendrograms}
\title{Title}
\usage{
plot_dendrogram(
  hcl.obj,
  method.dist = NULL,
  method.aggl = NULL,
  k = NULL,
  h = NULL,
  type = "rectangle",
  direction = "bt",
  branch.size = 1,
  display.labels = FALSE,
  labels.angle = 90,
  labels.hjust = 0,
  labels.nudge = 0.01,
  labels.size = 3,
  labels.vjust = 0.5,
  display.legend = TRUE,
  display.title = FALSE,
  clrp = "milo",
  clrp.adjust = NULL,
  ...
)

plot_dendrograms(
  hcl.obj,
  methods.dist = NULL,
  methods.aggl = NULL,
  k = NULL,
  h = NULL,
  type = "rectangle",
  direction = "bt",
  branch.size = 1,
  display.labels = FALSE,
  labels.angle = 90,
  labels.hjust = 0,
  labels.nudge = 0.01,
  labels.size = 3,
  labels.vjust = 0.5,
  display.legend = FALSE,
  display.title = TRUE,
  clrp = "milo",
  nrow = NULL,
  ncol = NULL,
  verbose = TRUE,
  return.grid = FALSE,
  ...
)
}
\arguments{
\item{...}{}

\item{fan}{}
}
\value{



}
\description{
Title
}
