% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-statistics-single.R
\name{plot_density}
\alias{plot_density}
\alias{plot_histogram}
\alias{plot_ridgeplot}
\title{Plot distribution and results of statistical tests}
\usage{
plot_density(
  df,
  variables = NULL,
  across = NULL,
  across.subset = NULL,
  relevel = TRUE,
  display.facets = TRUE,
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  clrp = "milo",
  clrp.adjust = NULL,
  pretty.names = TRUE,
  verbose = TRUE,
  ...
)

plot_histogram(
  df,
  variables = NULL,
  across = NULL,
  across.subset = NULL,
  relevel = TRUE,
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  clrp = "milo",
  clrp.adjust = NULL,
  pretty.names = TRUE,
  verbose = TRUE,
  ...
)

plot_ridgeplot(
  df,
  variables = NULL,
  across = NULL,
  across.subset = NULL,
  relevel = TRUE,
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  clrp = "milo",
  clrp.adjust = NULL,
  pretty.names = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{df}{A data.frame.}

\item{variables}{Character vector. Denotes the variables
of interest. If set to NULL all valid variables of input for \code{df} are
considered (as long as the total number does not exceeds the limit).

If there are more variables you are interested in than they are variables
you are not interested in specify those that you are not interested in prefixed
with an \emph{'-'}. Variables prefixed that way are discarded and the remaining are kept.}

\item{across}{Character value. Denotes the discrete variable in the data.frame
across which the variables of interest are to be analyzed or displayed.}

\item{across.subset}{Character vector. The groups of interest that the grouping variable
denoted in \code{across} contains.

If there are more groups you are interested in than they are groups you are not interested
in specify those that you are not interested in prefixed with an \emph{'-'}.
Variables prefixed that way are discarded and the remaining are kept.}

\item{relevel}{Logical value. If set to TRUE the input of \code{across.subset}
determines the new order in which the results are displayed.}

\item{display.facets}{Logical value. If set to TRUE the plot is split via
\code{ggplot2::facet_wrap()} such that each variable gets it's own subplot.}

\item{scales}{Given to \code{ggplot2::facet_wrap()}.}

\item{nrow}{Given to \code{ggplot2::facet_wrap()}.}

\item{ncol}{Given to \code{ggplot2::facet_wrap()}.}

\item{clrp}{Character value. The color panel of choice if \code{variable} is discrete
Run \code{confuns::all_colorpanels()} to see all valid input choices.

(Ignored if \code{variable} is numeric)}

\item{clrp.adjust}{Character vector or NULL. Adjusts the resulting named color vector.
Only named elements are kept. Elements of the color vector with the same names as in
\code{clrp.adjust} are replaced such that the elements of \code{clrp.adjust} are displayed by
the color their name indicates rather than the color the color panel would have assigned
to them.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Additional arguments given to the respective \code{ggplot2::geom_<plot_type>()}
function.}
}
\value{

}
\description{
Plot distribution and results of statistical tests
}
