% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive-statistics-plot.R
\name{plot_descriptive_statistics}
\alias{plot_descriptive_statistics}
\title{Plot descriptive statistics}
\usage{
plot_descriptive_statistics(
  df,
  variables = "all",
  across,
  across.subset = NULL,
  plot.type = "boxplot",
  binwidth,
  test.pairwise = "none",
  test.groupwise = "none",
  display.points = FALSE,
  pt.size = 1.2,
  pt.alpha = 0.8,
  pt.color = "black",
  n.obs = 100,
  shape.to = "none",
  ref.group,
  clrp = "milo",
  ...,
  verbose = TRUE
)
}
\arguments{
\item{df}{A data.frame that contains the grouping variable specified in \code{across}.}

\item{variables}{Character vector. The names of the numeric variables of interest.}

\item{across}{Character value. Denotes the discrete variable in the data.frame
across which the variables of interest are to be analyzed or displayed.}

\item{across.subset}{Character vector. The groups of interest that the grouping variable
denoted in \code{across} contains. If prefixed with an \emph{'-'} the groups are discarded.}

\item{plot.type}{Character value. One of \emph{'histogram', 'ridgeplot', 'boxplot',
'violinplot'} or \emph{'density'} to visualize the value distribution of those
variables specified via the \code{variables} argument.

If set to \emph{'boxplot'} or \emph{'violinplot'} and only one variable is
specified statistical test can be performed.}

\item{binwidth}{Numeric value. Only relevant if \code{plot_type} is set to \emph{'histogram'}.}

\item{test.pairwise}{Character value. One of \emph{'none', 't.test', 'wilcox.test'}.}

\item{test.groupwise}{Character value. One of \emph{'none', 'anova', 'kruskal.test'}.}

\item{display.points}{Logical. If set to TRUE the value distribution of \code{n.cells} is additionally
displayed by points.}

\item{n.obs}{Numeric value.}

\item{shape.to}{Character value. Denotes the discrete variable to which the shape of the
points are mapped. (Ignored if \code{display.points} is set to FALSE.)}

\item{ref.group}{Character value. Denotes the reference group for the statistical tests. Must
be one value of the variable specified in \code{across}.}

\item{...}{Additional arguments given to \code{ggplot2::facet_wrap()}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\description{
This function allows for a variety of different plots concerning
descriptive statistics.
}
