% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional-reduction.R
\name{plot_dim_red}
\alias{plot_dim_red}
\title{Title}
\usage{
plot_dim_red(
  dimred.obj,
  clr.aes = "fill",
  clr.by = NULL,
  clrp.adjust = NULL,
  pt.alpha = 0.9,
  pt.clr = "black",
  pt.clrp = "milo",
  pt.clrsp = "inferno",
  pt.fill = "black",
  pt.shape = 21,
  pt.size = 3,
  ncol = NULL,
  nrow = NULL,
  add.df = NA,
  add.ons = list(),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
