% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dot-plot.R
\name{plot_dot_plot_2d}
\alias{plot_dot_plot_2d}
\title{Plot 2d dot plot}
\usage{
plot_dot_plot_2d(
  df,
  x,
  y,
  alpha.by = NULL,
  alpha.trans = "identity",
  color.by = NULL,
  color.trans = "identity",
  shape.by = NULL,
  size.by = NULL,
  size.trans = "reverse",
  pt.alpha = 0.9,
  pt.color = "black",
  pt.clrp = "default",
  pt.clrsp = "plasma",
  pt.shape = 19,
  pt.size = 3,
  transform.with = NULL,
  arrange.y = FALSE,
  arrange.by = NULL,
  reverse.all = FALSE,
  reverse.within = FALSE,
  ...
)
}
\arguments{
\item{df}{A data.frame.}

\item{x, y}{Character value. Variable to be plotted on the respective axis.}

\item{arrange.y}{Logical. If TRUE the labels of \code{y} are arranged by
their appearances in the groups of \code{x}.}

\item{arrange.by}{Character value or NULL. If character, denotes the
numeric variable by which the labels of \code{y} are arranged within the
groups of \code{x}.}

\item{reverse.all}{Logical. If TRUE labels are displayed from bottom to top.
If FALSE, labels are displayed from top to bottom.}

\item{reverse.within}{Logical. If TRUE, labels are arranged within the
groups in a descending manner. If FALSE, labels are arranged within the
groups in an ascending manner. (regarding the numeric variables of \code{arrange.by}).}
}
\value{
A ggplot.
}
\description{
Plots a dot plot of two categorical variables. Numeric
values can be displayed via alpha, color and size.
}
