% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea-hypeR.R
\name{plot_gsea_dot}
\alias{plot_gsea_dot}
\alias{plot_gsea_dot.hyp}
\alias{plot_gsea_dot.list}
\alias{plot_gsea_dot.data.frame}
\title{Plot GSEA hypergeometric test}
\usage{
plot_gsea_dot(object, ...)

\method{plot_gsea_dot}{hyp}(
  object,
  n.gsets = 20,
  signif.val = "fdr",
  signif.threshold = 0.05,
  color.by = "fdr",
  size.by = "geneset",
  pt.size = 2,
  pt.color = "blue4",
  pt.clrsp = "plasma",
  remove = "MF.GO|RCTM|CC.GO|HM",
  replace = c("_", " "),
  ...
)

\method{plot_gsea_dot}{list}(
  object,
  n.gsets = 20,
  signif.val = "fdr",
  signif.threshold = 0.05,
  color.by = "fdr",
  size.by = "geneset",
  pt.size = 2,
  pt.color = "blue4",
  pt.clrsp = "plasma",
  remove = "MF.GO|RCTM|CC.GO|HM",
  replace = c("_", " "),
  ...
)

\method{plot_gsea_dot}{data.frame}(
  object,
  n.gsets = 20,
  signif.val = "fdr",
  signif.threshold = 0.05,
  color.by = "fdr",
  size.by = "geneset",
  pt.size = 2,
  pt.color = "blue4",
  pt.clrsp = "plasma",
  remove = "^.*_",
  replace = c("_", " "),
  ...
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{...}{Additional arguments given to \code{scale_color_add_on()}.}

\item{n.gsets}{Numeric value. Maximal number of gene sets whose
results are plotted. The first \code{n.sets} are included starting with the
one with the lowest significance value.}

\item{signif.val}{Character value. Either \emph{'pval'} or \emph{'fdr'}.}

\item{signif.threshold}{Numeric value. The maximum significance value a gene set
can have to be included.}

\item{color.by}{Character value or NULL. If character, specifies the
variable that is displayed by color.}

\item{size.by}{Character value or NULL. If character, the variable whose values
are displayed by size. Either \emph{'geneset'}, the number of genes the gene set
contains, or \emph{'overlap'}, the number of genes that overlapped between
the provided gene signature and the gene set.}

\item{pt.size}{Numeric values. Specifies size of points.}

\item{pt.color}{Character value. Specifies the color with which all points are displayed .}

\item{pt.clrsp}{Character value. Specfies the color spectrum to be used to represent
continuous values of numeric variables. Run \code{validColorSpectra()} to obtain
valid input options.}

\item{remove}{Character value or NULL. If character, regular expression given
to \code{pattern} of \code{stringr::str_remove_all()}. If character, regular expression given to
\code{pattern} of \code{stringr::str_replace_all()}. Used to adjust gene set names.}

\item{replace}{Character vector of length 2 or NULL. If character vector, two regular expressions.
First is given to argument \code{pattern} and second is given to argument \code{replacement}
of \code{stringr::str_replace_all()}. Used to adjust gene set names.}
}
\value{
A ggplot.
}
\description{
Plots p-value or FDR results of hyper geometrics tests.
}
