% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam-wrapper.R
\name{plot_medoid_barchart}
\alias{plot_medoid_barchart}
\title{Title}
\usage{
plot_medoid_barchart(
  pam.obj,
  k = NULL,
  metric.pam = NULL,
  cluster.prefix = "",
  cluster.subset = NULL,
  cluster.relevel = TRUE,
  variables.subset = NULL,
  variables.relevel = TRUE,
  display.medoid.name = FALSE,
  sep = ": ",
  clr = "black",
  clrp = "milo",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
