% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-statistics-single.R
\name{plot_mosaic}
\alias{plot_mosaic}
\title{Plot mosaic plot}
\usage{
plot_mosaic(df, x, fill.by, clrp = "default", clrp.adjust = NULL, ...)
}
\arguments{
\item{df}{A data.frame.}

\item{x}{Character value. The variable mapped to the x-axis.}

\item{fill.by}{Character value. The variable by which the mosaics
are filled.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp.adjust}{Named character vector or NULL. If character, it adjusts the
colors that are used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{...}{Additional arguments given to \code{ggmosaic::geom_mosaic()}}
}
\description{
Plots a mosaic plot.
}
