% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam-wrapper.R
\name{plot_pam_cluster_count}
\alias{plot_pam_cluster_count}
\title{Title}
\usage{
plot_pam_cluster_count(
  pam.obj,
  k = NULL,
  metric.pam = NULL,
  cluster.subset = NULL,
  cluster.relevel = TRUE,
  cluster.prefix = "",
  clr = "black",
  clrp = "milo",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
