% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam-clustering.R
\name{plot_pam_silinfo}
\alias{plot_pam_silinfo}
\title{Silhouette Width Info}
\usage{
plot_pam_silinfo(pam.obj)
}
\arguments{
\item{pam.obj}{A valid pam-object.}

\item{legend.position}{Character. Given to \code{ggplot2::theme()}.}
}
\description{
Plots the silhouette information of
the provided pam-object.
}
