% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-scatter.R
\name{plot_scatterplot}
\alias{plot_scatterplot}
\title{Title}
\usage{
plot_scatterplot(
  df,
  x,
  y,
  across = NULL,
  across.subset = NULL,
  relevel = TRUE,
  ncol = NULL,
  nrow = NULL,
  scales = "fixed",
  space = "fixed",
  pt_alpha = 0.9,
  pt_clr = "black",
  pt_fill = "black",
  pt_shape = 21,
  pt_size = 2,
  smooth_clr = "blue",
  smooth_method = NULL,
  smooth_se = FALSE
)
}
\arguments{
\item{smooth_se}{}
}
\value{

}
\description{
Title
}
