% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans-wrapper.R
\name{plot_screeplot}
\alias{plot_screeplot}
\title{Title}
\usage{
plot_screeplot(
  kmeans.obj,
  methods.kmeans = NULL,
  clr = "steelblue",
  display.cols = TRUE,
  display.line = TRUE,
  display.points = TRUE
)
}
\arguments{
\item{display.points}{}
}
\value{

}
\description{
Title
}
