% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam-wrapper.R
\name{plot_silhouette_widths}
\alias{plot_silhouette_widths}
\title{Title}
\usage{
plot_silhouette_widths(
  pam.obj,
  metric.pam = NULL,
  k = NULL,
  clrp = "milo",
  ncol = NULL,
  nrow = NULL,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
