% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-statistics-interactive.R
\name{plot_statistics_interactive}
\alias{plot_statistics_interactive}
\title{Opens interactive application}
\usage{
plot_statistics_interactive(df, n.across.subset = 10)
}
\arguments{
\item{df}{A data.frame.}
}
\description{
Opens interactive application
}
