% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-statistics-single.R
\name{plot_violin}
\alias{plot_violin}
\alias{plot_boxplot}
\title{Plot distribution and results of statistical tests}
\usage{
plot_violin(
  df,
  variables = NULL,
  across = NULL,
  across.subset = NULL,
  relevel = TRUE,
  test.pairwise = NULL,
  test.groupwise = NULL,
  ref.group = NULL,
  step.increase = 0.1,
  vjust = 0,
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  display.points = FALSE,
  pt.alpha = 0.8,
  pt.color = "black",
  pt.num = 100,
  pt.size = 1.5,
  pt.shape = 19,
  clrp = "milo",
  clrp.adjust = NULL,
  pretty.names = TRUE,
  verbose = TRUE,
  ...
)

plot_boxplot(
  df,
  variables = NULL,
  across = NULL,
  across.subset = NULL,
  relevel = TRUE,
  test.pairwise = NULL,
  test.groupwise = NULL,
  ref.group = NULL,
  step.increase = 0.1,
  vjust = 0,
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  display.points = FALSE,
  pt.alpha = 0.8,
  pt.color = "black",
  pt.size = 1.5,
  pt.num = 100,
  pt.shape = 19,
  clrp = "milo",
  clrp.adjust = NULL,
  pretty.names = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{df}{A data.frame.}

\item{variables}{Character vector. Denotes the variables
of interest. If set to NULL all valid variables of input for \code{df} are
considered (as long as the total number does not exceeds the limit).

If there are more variables you are interested in than they are variables
you are not interested in specify those that you are not interested in prefixed
with an \emph{'-'}. Variables prefixed that way are discarded and the remaining are kept.}

\item{across}{Character value. Denotes the discrete variable in the data.frame
across which the variables of interest are to be analyzed or displayed.}

\item{across.subset}{Character vector. The groups of interest that the grouping variable
denoted in \code{across} contains.

If there are more groups you are interested in than they are groups you are not interested
in specify those that you are not interested in prefixed with an \emph{'-'}.
Variables prefixed that way are discarded and the remaining are kept.}

\item{relevel}{Logical value. If set to TRUE the input of \code{across.subset}
determines the new order in which the results are displayed.}

\item{test.pairwise}{Character value or NULL. If character, one of \emph{'none', 't.test', 'wilcox.test'}.}

\item{test.groupwise}{Character value or NULL. If character, one of \emph{'none', 'anova', 'kruskal.test'}.}

\item{ref.group}{Character value. Denotes the reference group for the statistical tests. Must
be one value of the variable specified in \code{across}.}

\item{step.increase}{Numeric value. Denotes the increase in fraction of total
height for every additional comparison to minimize overlap.}

\item{vjust}{Numeric value. Denotes the relative, vertical position of the results of
the test denoted in \code{test.groupwise}. Negative input highers, positive
input lowers the position.}

\item{scales}{Given to \code{ggplot2::facet_wrap()}.}

\item{nrow}{Given to \code{ggplot2::facet_wrap()}.}

\item{ncol}{Given to \code{ggplot2::facet_wrap()}.}

\item{pt.alpha}{Numeric value. Denotes the transperancy of the points.}

\item{pt.color}{Character value. Denotes the color with
which the points are displayed .}

\item{pt.num}{Numeric value. Regulates the number of points (sample size)
to prevent overplotting.}

\item{pt.size}{Numeric values. Denote size of the points.}

\item{pt.shape}{Numeric or character value. If numeric, the respective
shape is taken for all points. If character, the respective variable is
mapped to the shape-aesthetic.}

\item{clrp}{Character value. The color panel of choice if \code{variable} is discrete
Run \code{confuns::all_colorpanels()} to see all valid input choices.

(Ignored if \code{variable} is numeric)}

\item{clrp.adjust}{Character vector or NULL. Adjusts the resulting named color vector.
Only named elements are kept. Elements of the color vector with the same names as in
\code{clrp.adjust} are replaced such that the elements of \code{clrp.adjust} are displayed by
the color their name indicates rather than the color the color panel would have assigned
to them.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Additional arguments given to the respective \code{ggplot2::geom_<plot_type>()}
function.}
}
\value{

}
\description{
Plot distribution and results of statistical tests
}
