% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Analysis.R
\name{plotBoxplot}
\alias{plotBoxplot}
\alias{plotBoxplot,Analysis-method}
\alias{plotViolinplot,Analysis-method}
\title{Plot boxplot}
\usage{
plotBoxplot(object, ...)

\S4method{plotBoxplot}{Analysis}(
  object,
  variables,
  phase = NULL,
  across = NULL,
  across_subset = NULL,
  relevel = TRUE,
  clrp = "milo",
  clrp_adjust = NULL,
  test_groupwise = NULL,
  test_pairwise = NULL,
  ref_group = NULL,
  step_increase = 0.01,
  vjust = 0,
  display_facets = TRUE,
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  display_points = FALSE,
  pt_alpha = 0.8,
  pt_clr = "black",
  pt_num = 100,
  pt_size = 1.25,
  pt_shape = 21,
  verbose = TRUE,
  ...
)

\S4method{plotViolinplot}{Analysis}(
  object,
  variables,
  phase = NULL,
  across = NULL,
  across_subset = NULL,
  relevel = TRUE,
  clrp = "milo",
  clrp_adjust = NULL,
  test_groupwise = NULL,
  test_pairwise = NULL,
  ref_group = NULL,
  step_increase = 0.01,
  vjust = 0,
  display_facets = TRUE,
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  display_points = FALSE,
  pt_alpha = 0.8,
  pt_clr = "black",
  pt_num = 100,
  pt_size = 1.25,
  pt_shape = 21,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{variables}{Character vector. Specifies the variables
of interest. If set to NULL all valid variables of the input data.frame are
considered (as long as the total number does not exceeds the limit).

If there are more variables you are interested in than they are variables
you are not interested in specify those that you are not interested in prefixed
with an \emph{'-'}. Variables prefixed that way are discarded and the remaining are kept.}

\item{across}{Character value. Specifies the discrete variable in the data.frame
across which the variables of interest are to be analyzed or displayed.}

\item{relevel}{Logical value. If set to TRUE the input order of \code{across.subset} and
\code{variables} determines the order in which the groups of interest are displayed.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{scales, nrow, ncol}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A ggplot.
}
\description{
Plots a boxplot
}
