% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Correlation.R
\name{plotCorrplot}
\alias{plotCorrplot}
\alias{plotCorrplot,Correlation-method}
\title{Plot correlation plot}
\usage{
plotCorrplot(object, ...)

\S4method{plotCorrplot}{Correlation}(
  object,
  method_corr = "pearson",
  across = NULL,
  across_subset = NULL,
  variables_subset = NULL,
  pval_threshold = NULL,
  type = "lower",
  diagonal = TRUE,
  color_low = "darkred",
  color_high = "steelblue",
  shape = "tile",
  size_by_corr = TRUE,
  size_max = 15,
  display_values = TRUE,
  values_alpha = 0.9,
  values_color = "black",
  values_digits = 2,
  values_size = 4,
  display_grid = TRUE,
  grid_color = "black",
  grid_size = 0.5,
  nrow = NULL,
  ncol = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{method_corr}{Character value. Specifies the correlation
method of interest. Use \code{validMethodsCorrelation()} to obtain all valid
input options.}

\item{across}{Character value. Specifies the discrete variable in the data.frame
across which the variables of interest are to be analyzed or displayed.}

\item{pval_threshold}{Numeric value or NULL. If numeric, denotes the maximum p-value to be considered as
significant. Must no be higher than 0.05.}

\item{nrow}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{ncol}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A ggplot.
}
\description{
Plot correlation plot
}
