% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R
\name{plotRidgeplot}
\alias{plotRidgeplot}
\title{Plot ridgeplot}
\usage{
plotRidgeplot(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
A ggplot.
}
\description{
Plots a ridgeplot
}
