% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-AnalysisAspect.R
\name{plotScatterplot}
\alias{plotScatterplot}
\alias{plotScatterplot,AnalysisAspect-method}
\title{Plot a scatterplot}
\usage{
plotScatterplot(object, ...)

\S4method{plotScatterplot}{AnalysisAspect}(
  object,
  x,
  y,
  across = NULL,
  across_subset = NULL,
  relevel = TRUE,
  ncol = NULL,
  nrow = NULL,
  scales = "fixed",
  space = "fixed",
  pt_alpha = 0.9,
  pt_color = "black",
  pt_clrp = "milo",
  pt_fill = "black",
  pt_shape = 19,
  pt_size = 1.5,
  color_aes = "color",
  color_by = NULL,
  clrp = "milo",
  clrp_adjust = NULL,
  clrsp = "inferno",
  display_smooth = FALSE,
  smooth_alpha = 0.9,
  smooth_color = "blue",
  smooth_method = "lm",
  smooth_se = FALSE,
  smooth_size = 1,
  display_corr = FALSE,
  corr_method = "pearson",
  corr_p_min = 5e-05,
  corr_pos_x = NULL,
  corr_pos_y = NULL,
  corr_text_sep = "\\n",
  corr_text_size = 1,
  ...
)
}
\value{
A ggplot.
}
\description{
Plots two numeric variables on the x- and y-axis.
}
