% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Clustering.R,
%   R/S4-Analysis.R
\name{plotScreeplot}
\alias{plotScreeplot}
\alias{plotScreeplot,Clustering-method}
\alias{plotScreeplot,Analysis-method}
\title{Plot a screeplot}
\usage{
plotScreeplot(object, ...)

\S4method{plotScreeplot}{Clustering}(
  object,
  methods_kmeans,
  ks = NULL,
  color = "steelblue",
  display_cols = TRUE,
  col_alpha = 0.9,
  col_color = "black",
  col_fill = "steelblue",
  display_line = TRUE,
  line_alpha = 0.9,
  line_color = "black",
  line_size = 1.5,
  display_points = TRUE,
  pt_alpha = 0.9,
  pt_color = "black",
  pt_size = 4.5
)

\S4method{plotScreeplot}{Analysis}(
  object,
  methods_kmeans = "Hartigan-Wong",
  ks = NULL,
  display_cols = TRUE,
  col_alpha = 0.9,
  col_color = "black",
  col_fill = "steelblue",
  display_line = TRUE,
  line_alpha = 0.9,
  line_color = "black",
  line_size = 1.5,
  display_points = TRUE,
  pt_alpha = 0.9,
  pt_color = "black",
  pt_size = 4.5
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{ks}{Numeric vector. Denotes all options for k-clusters. Values <1 are discarded.
Is converted into an integer vector.}
}
\value{
A ggplot.
}
\description{
Plots quality of clustering results
suggested by kmeans with a screeplot.
}
\details{
If \code{ks} and \code{methods_kmeans} are specified only
the resulting combinations are included. Else all found combinations
are included.
}
