% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-wrapper.R
\name{prepare_cluster_data}
\alias{prepare_cluster_data}
\title{Title}
\usage{
prepare_cluster_data(
  data,
  make.names = TRUE,
  names.prefix = NULL,
  names.sep = "-"
)
}
\arguments{
\item{names.sep}{}
}
\value{

}
\description{
Title
}
