% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_infos.R
\name{print_gene_info}
\alias{print_gene_info}
\title{Print gene information}
\usage{
print_gene_info(genes, check = TRUE)
}
\arguments{
\item{genes}{Character vector of gene names.}

\item{check}{Logical value. If TRUE, input is checked for availability. If
FALSE, unknown elements are silently dropped.}
}
\value{
Invisible TRUE. Texts are immediately printed using \code{base::writeLines()}.
}
\description{
Prints human readable summary texts of genes in the console.
}
\details{
See documentation for \code{?gene_info_df} to read about how the source
data.frame has been generated.
}
\examples{

 # genes are checked for validity by default
 print_gene_info(genes = c("GFAP", "MAG", "OLIG1", "XYZ"))

 # set check = FALSE to drop unkonwn genes silently
 print_gene_info(genes = c("GFAP", "MAG", "OLIG1", "XYZ"), check = FALSE)


}
