% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-project.R
\name{proj_py_install_pkgs}
\alias{proj_py_install_pkgs}
\title{Install Python Packages into a Virtual Environment}
\usage{
proj_py_install_pkgs(pkgs, venv = "venv_main", ...)
}
\arguments{
\item{pkgs}{A vector of package names to install.}

\item{venv}{The name of the virtual environment where packages will be installed.
Default is 'venv_main'.}

\item{...}{Additional arguments passed to \code{virtualenv_install}.}
}
\value{
NULL This function attempts to install packages and handles errors silently,
returning no value.
}
\description{
This function installs the specified Python packages into the given virtual environment.
It attempts to install each package and catches any errors during the installation process.
}
\examples{
proj_py_install_pkgs(c("numpy", "pandas"), "venv_main")

}
