% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-project.R
\name{proj_py_pkg_installed}
\alias{proj_py_pkg_installed}
\title{Check if a Python Package is Installed}
\usage{
proj_py_pkg_installed(pkg, version = NULL)
}
\arguments{
\item{pkg}{The name of the package to check.}

\item{version}{Optional specific version to check for. If provided, checks for the package
and the specific version.}
}
\value{
Logical indicating whether the specified package (and version) is installed.
}
\description{
Checks if a specified package, and optionally a specific version, is installed in the active
Python environment. The function retrieves a list of installed packages and searches for the specified package.
}
\examples{
proj_py_pkg_installed("numpy")
proj_py_pkg_installed("pandas", "1.1.5")

}
