% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-project.R
\name{proj_py_pkgs}
\alias{proj_py_pkgs}
\title{List Installed Python Packages}
\usage{
proj_py_pkgs()
}
\value{
A tibble with two columns: 'pkg' for package names and 'version' for package versions.
}
\description{
Retrieves a list of installed Python packages in the active virtual environment using \verb{pip list}.
The output is returned as a tibble with columns for package names and versions.
}
\examples{
proj_py_pkgs()

}
