% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-project.R
\name{proj_py_use_venv}
\alias{proj_py_use_venv}
\title{Use a Specific Python Virtual Environment}
\usage{
proj_py_use_venv(venv = "venv_main")
}
\arguments{
\item{venv}{The name of the virtual environment to activate. Default is 'venv_main'.}
}
\value{
NULL This function is used for its side effect of activating a Python virtual
environment and does not return any value.
}
\description{
This function sets the specified Python virtual environment as active for the current R session.
It checks if the virtual environment exists at the specified path; if not, it throws an error.
}
\examples{
proj_py_use_venv("venv_main")

}
