% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-algebra.R
\name{pythagoras_theorem}
\alias{pythagoras_theorem}
\title{Pythagoras theorem}
\usage{
pythagoras_theorem(a = NA, b = NA, c = NA)
}
\arguments{
\item{a, b, c}{Numeric values. The one that is \code{NA} is computed.}
}
\value{
Numeric value.
}
\description{
Computes the missing part of the variables of the theorem
of pythagoras \emph{a^2 x b^2 = c^2}.
}
