% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-algebra.R
\name{rad2degr}
\alias{rad2degr}
\title{Convert radian to degree}
\usage{
rad2degr(rad, d = Inf)
}
\arguments{
\item{rad}{Numeric vector of values in radian. Values must frange from 0 to \code{2*pi}.}

\item{d}{Number of digits to which the output values are rounded.}
}
\value{
Numeric vector of the same length.
}
\description{
Converts radians to degrees.
}
