% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-algebra.R
\name{radius}
\alias{radius}
\title{Compute circle radius}
\usage{
radius(circ, digits = Inf)
}
\arguments{
\item{circ}{Numeric vector. The circumference/circumferences.}
}
\value{
Numeric vector of the same length.
}
\description{
Computes the radius of a circle based on the circumference
}
