% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-wrapper.R
\name{rename_numeric_vars}
\alias{rename_numeric_vars}
\alias{rename_numeric_vars_with}
\title{Rename numeric variables of corr_conv objects}
\usage{
rename_numeric_vars(corr.obj, ..., rename.data = TRUE)

rename_numeric_vars_with(corr.obj, ..., rename.data = TRUE)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Rename numeric variables of corr_conv objects
}
