% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-leverage.R
\name{rename_savely}
\alias{rename_savely}
\title{Save wrapper around dplyr::rename()}
\usage{
rename_savely(df, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Save wrapper around dplyr::rename()
}
