% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-AnalysisAspect.R
\name{renameKeyVariable}
\alias{renameKeyVariable}
\alias{renameKeyVariable,AnalysisAspect-method}
\title{Rename key variable}
\usage{
renameKeyVariable(object, ...)

\S4method{renameKeyVariable}{AnalysisAspect}(object, new_key_name)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
The input object.
}
\description{
Renames the variable that is used to identify every
observation uniquely.
}
