% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-handling.R
\name{rgx_lookahead}
\alias{rgx_lookahead}
\title{Create lookahead/-behind regex}
\usage{
rgx_lookahead(pattern, match = ".*", negate = FALSE)
}
\arguments{
\item{pattern}{Character value. The pattern before/behind which the subsequent
function looks.}

\item{match}{Character value. The regex that must match the string before/behind
the input of \code{pattern}.}

\item{negate}{Logical value. If TRUE, function matches the part of input that
does \bold{not} match input for argument \code{match}.}
}
\description{
Creates a lookahead/-behind regex. See details for more.
}
