% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-manipulation.R
\name{scollapse}
\alias{scollapse}
\title{Collapse with glue::glue_collapse}
\usage{
scollapse(string, sep = "', '", width = Inf, last = "' and '")
}
\arguments{
\item{string}{Character vector to be collapsed.}
}
\description{
Collapse with glue::glue_collapse
}
