% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_infos.R
\name{search_gene_names}
\alias{search_gene_names}
\title{Search for genes}
\usage{
search_gene_names(catchphrases, test = "any", return_var = "symbol", ...)
}
\arguments{
\item{catchphrases}{Character vector of catchphrases to look for.}

\item{test}{Character value. Determines how the input of \code{catchphrases}
is used to skim the genes. If \emph{'any'} genes are kept if at least
one of the catchphrases appears in the summary. If \emph{'all'}, genes are kept if all
of the catchphrases appear in the summary. If \emph{'none'}, genes are kept
if none of the catchphrases appear in the summary.}

\item{return_var}{The variable of the \code{gene_info_df} that is returned
as a vector. Defaults to \emph{symbol}.}

\item{...}{Logical tests given to \code{dplyr::filter()} that can be used to
prefilter the data.frame before going through the summary texts.}
}
\value{
Character vector of gene names.
}
\description{
Goes through the summary text of genes and returns
gene names that fit the requirements.
}
