% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame-handling.R
\name{select_columns}
\alias{select_columns}
\title{Wrapper around dplyr::select()}
\usage{
select_columns(
  df,
  keep = NULL,
  contains = NULL,
  matches = NULL,
  starts.with = NULL,
  ends.with = NULL,
  negate = FALSE,
  return = "tibble",
  fdb.fn = "stop"
)
}
\arguments{
\item{df}{A data.frame.}

\item{keep}{Columns that have to be kept.}

\item{contains, matches, starts.with, ends.with}{Character or NULL. If character
given to the respective tidyselect function as input for argument \code{match}.}

\item{negate}{Logical value. If set to TRUE negates the regex input of the tidyselect
input.}
}
\value{
Selected data.frame.
}
\description{
Wrapper around dplyr::select()
}
