% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes-and-spectra.R
\docType{data}
\name{sequential_single_hue}
\alias{sequential_single_hue}
\alias{sequential_multi_hue}
\alias{diverging}
\title{Color palette names}
\format{
An object of class \code{character} of length 10.

An object of class \code{character} of length 50.

An object of class \code{character} of length 16.
}
\usage{
sequential_single_hue

sequential_multi_hue

diverging
}
\description{
Color palette names
}
\keyword{datasets}
