% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-wrapper.R
\name{set_corr_default}
\alias{set_corr_default}
\title{Title}
\usage{
set_corr_default(corr.obj, method.corr = NA, directory = NA)
}
\arguments{
\item{directory}{}
}
\value{

}
\description{
Title
}
