% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R
\name{set_data_hlpr}
\alias{set_data_hlpr}
\title{Set data helper}
\usage{
set_data_hlpr(
  object,
  data,
  key.name,
  key.prefix,
  meta.names,
  slot.data = "data",
  slot.key.name = "key_name",
  slot.meta = "meta",
  verbose = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object.
}
\description{
Helper for generic function \code{setData()}.
}
