% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust-wrapper.R
\name{set_hclust_default}
\alias{set_hclust_default}
\title{Title}
\usage{
set_hclust_default(
  hcl.obj,
  method.aggl = NA,
  method.dist = NA,
  directory = NA,
  verbose = FALSE
)
}
\arguments{
\item{method.dist}{}
}
\value{

}
\description{
Title
}
