% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam-wrapper.R
\name{set_pam_default}
\alias{set_pam_default}
\title{Title}
\usage{
set_pam_default(
  pam.obj,
  as.dist = NULL,
  directory = NULL,
  k = NULL,
  metric.pam = NULL,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
