% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-project.R
\name{set_up_project}
\alias{set_up_project}
\title{Set Up Project Directories with Optional Python Environment}
\usage{
set_up_project(
  project_name,
  python = TRUE,
  python_version = NULL,
  python_pkgs = NULL
)
}
\arguments{
\item{project_name}{The name of the project which should match the end of the
current working directory's path.}

\item{python}{Logical, if \code{TRUE}, sets up a Python virtual environment and
script directories. Defaults to \code{TRUE}.}

\item{python_version}{The version of Python to use for the virtual environment.
If \code{NULL}, uses the system's default Python version.}

\item{python_pkgs}{A character vector of additional Python packages to install
in the virtual environment, beyond the defaults (matplotlib, numpy, pandas).}
}
\value{
NULL This function is used for its side effect of creating directories
and optionally a Python virtual environment. It does not return any value.
}
\description{
Initializes a new project directory structure in the current working directory.
Verifies if the working directory's name matches the provided project name.
If the names do not match, the execution is stopped with an error message.
Optionally sets up a Python virtual environment and installs specified packages.
}
\details{
The directories created are as follows:
\itemize{
\item{\code{data}: Main directory for all data-related storage.
\itemize{
\item{\code{raw}: For storing raw, unmodified data.}
\item{\code{proc}: For storing processed data, which is derived from raw data.}
\item{\code{out}: For storing output data, such as results or final datasets.}
}
}
\item{\code{docs}: For documentation files related to the project.}
\item{\code{figures}: For figures (.ai, .svg, ...).}
\item{\code{plots}: For storing graphical representations generated from the data.}
\item{\code{R_markdown}: For storing R Markdown documents used for reports or presentations.}
\item{\code{R_scripts}: For storing R scripts that are part of the project.}
\item{\code{py_scripts} and \code{py_venv}: Created only if \code{python} is \code{TRUE}. The \code{py_venv} directory
contains a Python virtual environment named \code{venv_main} with specified Python packages installed.}
}
}
\examples{
# Set up a project named "MyProject" with a Python virtual environment
set_up_project("MyProject", python = TRUE, python_version = "3.8", python_pkgs = c("scipy", "seaborn"))

# Set up a project without a Python environment
set_up_project("MyProject", python = FALSE)

}
