% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-Analysis.R
\name{setAnalysisAspect}
\alias{setAnalysisAspect}
\alias{setAnalysisAspect,Analysis-method}
\title{Set objects of class \code{AnalysisAspect}}
\usage{
setAnalysisAspect(object, ...)

\S4method{setAnalysisAspect}{Analysis}(object, aa)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{aa}{An object of class \code{AnalysisAspect}.}
}
\value{
The input object.
}
\description{
Sets objects of class \code{AnalysisAspect} and removes
content of slots @data, @data_scaled and @meta.
}
