% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-AnalysisAspect.R
\name{setData}
\alias{setData}
\alias{setData,AnalysisAspect-method}
\title{Set data and key variables}
\usage{
setData(object, ...)

\S4method{setData}{AnalysisAspect}(
  object,
  data,
  key_name = NULL,
  key_prefix = "id",
  meta_names = character(0),
  verbose = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{key_name}{Character value or NULL. Denotes the variable that is used
to identify each observation uniquely.}

\item{key_prefix}{Character value. The string with which
the IDs in the newly constructed key variable are prefixed if \code{key_name} is
NULL.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A data.frame with subclass \code{tibble}.
}
\description{
Sets the objects slot @data and slot @key_name and makes
sure that the key variable exists and is valid.
}
\details{
If \code{key_name} is NULL the key variable is constructed
as a combination of the input data.frames rownames and the string
provided with \code{key_prefix}. If the the datas rownames are NULL
or contain \emph{""} the rownumbers are used instead. The newly constructed
key variable is named \emph{data_ids}.
}
