% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-Analysis.R
\name{setInstruction}
\alias{setInstruction}
\alias{setInstruction,Analysis-method}
\title{Set instruction}
\usage{
setInstruction(object, ...)

\S4method{setInstruction}{Analysis}(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
The input object.
}
\description{
Sets instructions to adjust automatic behaviour.
}
