% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-project.R
\name{setup_project}
\alias{setup_project}
\title{Set up the project folder}
\usage{
setup_project(project.name)
}
\arguments{
\item{project.name}{The projects name as a character string to validate
the current working directory.}
}
\description{
Creates input-, meta-, output- and graphics directories relative
to the working directory.
}
