% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-wrapper-shiny.R
\name{shiny_fdb}
\alias{shiny_fdb}
\title{Show shiny - notifications}
\usage{
shiny_fdb(in.shiny, ui, type = "message", ...)
}
\arguments{
\item{in.shiny}{Logical value.}

\item{ui}{Given to \code{shiny::showNotification()}.}

\item{type}{Given to \code{shiny::showNotification()}.}

\item{...}{More arguments given \code{shiny::showNotification()}.}
}
\value{
A shiny notification.
}
\description{
Show shiny - notifications
}
