% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-colors.R
\name{show_colors}
\alias{show_colors}
\title{Visualize colorpalettes and -spectra}
\usage{
show_colors(input, n = 20, title_size = 10)
}
\arguments{
\item{input}{Character vector of colorpalettes and -spectra.}

\item{n}{Numeric. Indicates how many colors to be used.}

\item{title_size}{Numeric. The title size.}
}
\value{
A ggplot.
}
\description{
Visualizes colorpalettes and -spectra as referred
to by arguments \code{clrp} and \code{clrsp}.
}
