% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_and_load.R
\name{source_functions}
\alias{source_functions}
\title{Load in all projects}
\usage{
source_functions(folder = "R", recursive = FALSE, verbose = TRUE)
}
\arguments{
\item{folder}{The absolute or relative to the folder of interest.}

\item{recursive}{Logical. If set to TRUE the listing recurs into subsequent
folders.}

\item{verbose}{Logical. If set to TRUE informative messages will be printed.

(Warnings will always be printed.)}
}
\value{
Sourced in functions as a side effect and an invisible TRUE.
}
\description{
Sources all R-scripts from folder 'R' that end with
..._function.R or ..._functions.R.
}
