% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-statistics-add-ons.R
\name{statistics_geom_jitter}
\alias{statistics_geom_jitter}
\title{Geom jitter add on}
\usage{
statistics_geom_jitter(
  df_shifted,
  across,
  aes_x,
  display.points,
  pt.alpha,
  pt.color,
  pt.num,
  pt.shape,
  pt.size
)
}
\description{
Geom jitter add on
}
