% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-statistics-add-ons.R
\name{statistics_tests}
\alias{statistics_tests}
\title{Stat compare means}
\usage{
statistics_tests(
  df_shifted,
  across,
  ref.group,
  test.pairwise,
  test.groupwise,
  step.increase,
  vjust
)
}
\description{
Stat compare means
}
