% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-handling.R
\name{str_extract_before}
\alias{str_extract_before}
\alias{str_extract_after}
\title{Extract string content with lookarounds}
\usage{
str_extract_before(
  string,
  pattern,
  match = "^.*",
  negate = FALSE,
  cut.right = TRUE,
  cut.left = TRUE
)

str_extract_after(
  string,
  pattern,
  match = ".*$",
  negate = FALSE,
  cut.right = TRUE,
  cut.left = TRUE
)
}
\arguments{
\item{string}{Character value.}

\item{pattern}{The regular expression that matches the part of the string
before/after which is looked. (Not included in the output.)}

\item{match}{The regular expression that is matched to the part of the string
before/after \code{pattern}.}

\item{cut.left, cut.right}{Logical value. If TRUE, empty space of the remaining string
is removed on the left/right side. Defaults to TRUE.}
}
\value{
Character value.
}
\description{
Extracts content of strings that appear before or after a pattern.
}
