% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-handling.R
\name{str_suggest_vec}
\alias{str_suggest_vec}
\alias{str_suggest_list}
\title{Obtain string suggestions}
\usage{
str_suggest_vec(string, pool, max.dist = 1, n.try = 0, n.top = 5)

str_suggest_list(string, pool, max.dist = 1, n.try = 0, n.top = 5)
}
\arguments{
\item{string}{Input string. \code{str_suggest_vec()} takes a single
character value of length 1. \code{str_suggest_list()} takes a character
vector of arbitrary length.}

\item{pool}{Character vector that contains all valid options.}

\item{max.dist}{Numeric value. The maximum distance a string of the
pool may have to be included in the suggestions.}

\item{n.try}{Numeric value. The number of times the search for suggestions
is repeated with \code{max.dist+1}. Defaults to 0.}

\item{n.top}{Numeric value. The n best suggestions in terms of string
distance to input string that are returned.}

\item{out.fail}{The output value if no suggestions can be made.}
}
\value{
\code{str_suggest_vec()} returns a character vector of suggestions. Is
of length 0 if no suggestions can be made. \code{str_suggest_list()} returns
a named list of length equal to length of \code{string}. Each slot carries
suggestions of one element of the input for \code{string}.
}
\description{
Compares input string against a pool of options and returns
the values of the pool that are similar to the input string.
}
