% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap-related.R
\name{umap_add_layout}
\alias{umap_add_layout}
\title{UMAP calculation}
\usage{
umap_add_layout(df, scale = FALSE)
}
\arguments{
\item{df}{A data.frame.}

\item{scale}{Logical. If set to TRUE all numeric variables will be scaled
with \code{base::scale()} before the call to \code{umap::umap()} happens.}
}
\value{
The input data.frame specified in \code{df} with two additional
numeric variables named \emph{umap1} and \emph{umap2}.
}
\description{
Calculates and adds the umap-dimensions to the specified
data.frame.
}
\examples{
umap_add_layout(df = mtcars, scale = TRUE)
}
