% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{unfactor}
\alias{unfactor}
\title{Wrapper around unfactor()}
\usage{
unfactor(input, ...)
}
\arguments{
\item{input}{A vector.}

\item{...}{Additional arguments given to s4vctrs::unfactor().}
}
\description{
If input is a factor it will be returned as
a character vector.
}
