% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid-misc.R
\name{validColorPalettes}
\alias{validColorPalettes}
\alias{validColorSpectra}
\title{Valid color inputs}
\usage{
validColorPalettes()

validColorSpectra()
}
\value{
Character vector or list of character vectors.
}
\description{
Functions that return the valid input options
for arguments that take predefined color palettes or color spectra.
}
