% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-classes.R
\name{variable_classes}
\alias{variable_classes}
\title{Names with classes}
\usage{
variable_classes(data)
}
\arguments{
\item{data}{A data.frame or a list.}
}
\value{
A named character vector.
}
\description{
Takes either a list or a data.frame and returns
a named vector whereby the elements of the vector are the names
of the \code{data}-variables and the names of the vectors elements
are the class of that particular variable.
}
\examples{
variable_classes(data = mtcars)

}
