% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-classes.R
\name{variable_classes2}
\alias{variable_classes2}
\title{Classes with names}
\usage{
variable_classes2(data)
}
\arguments{
\item{data}{A data.frame or a list.}
}
\value{
A named character vector.
\itemize{
 \item{ \strong{Vector-elements:} The class of every variable of \code{data}}.
 \item{ \strong{Vector-names:} The name of every variable of \code{data}}.
 }
}
\description{
Get information about the inputs
variable names and classes.
}
\examples{
variable_classes2(data = mtcars)

}
