% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-leverage.R
\name{vredefine}
\alias{vredefine}
\alias{vredefine_with}
\alias{vrename}
\alias{vrename_with}
\title{Rename vector in tidyverse style}
\usage{
vredefine(input, ..., safely = TRUE)

vredefine_with(input, ...)

vrename(input, ..., safely = TRUE)

vrename_with(input, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Rename vector in tidyverse style
}
