% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap-strings.R
\name{wrap_strings}
\alias{wrap_strings}
\title{Wrap all elements of a character vector}
\usage{
wrap_strings(input, wrap.in = character())
}
\arguments{
\item{input}{An object that contains character strings.}

\item{wrap.in}{A character vector of length one or two.

If of length one: Pre- and Suffix of the results will be the same.

If of length two: \code{wrap.in[1]} will be the prefix. and \code{wrap.in[1]}
the suffix.}
}
\value{
The provided R-object. All elements of the provided character vector
are wrapped with a pre- and suffix. If \code{input} is not of type vector the
return value from \code{base::names(\code{input})} will be wrapped accordingly.
}
\description{
Wrap all elements of a character vector
}
