% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R
\name{addClusterVarsHclust}
\alias{addClusterVarsHclust}
\title{Add hierarchical cluster variables}
\usage{
addClusterVarsHclust(object, ...)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\description{
Adds hierarchical cluster results in form of grouping variables
to the data of the input object.
}
\details{
Iterates over all possible combinations of the arguments \code{ks}, \code{hs},
\code{methods_dist} and \code{methods_aggl} and adds the corresponding grouping
variables to the objects data. Cluster variables are named corresponding to the input
of arguments \code{naming_k} and \code{naming_h}.
}
