% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-Correlation.R
\name{adjust_corr_mtr}
\alias{adjust_corr_mtr}
\title{Adjust correlation matrix}
\usage{
adjust_corr_mtr(mtr, type = "complete", diagonal = TRUE)
}
\arguments{
\item{mtr}{Correlation or p-value matrix.}

\item{diagonal}{Logical value.  Indicates if the diagonal values of the
matrices are kept ( = TRUE) or set to NA (= FALSE). Ignored if \code{type}
= \emph{'complete'}.}
}
\value{
Adjusted input matrix.
}
\description{

}
