% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-Correlation.R
\name{adjust_corr_mtr}
\alias{adjust_corr_mtr}
\title{Adjust correlation matrix}
\usage{
adjust_corr_mtr(mtr, type = "complete", diagonal = TRUE)
}
\arguments{
\item{mtr}{Correlation or p-value matrix.}

\item{type}{Character value. Denotes how the underlying correlation matrix is
handled. Three options:

\itemize{
\item{\emph{'complete'}:}{ The matrix stays as is.},
\item{\emph{'lower'}:}{ The part below the diagonal is used. The upper part is set to NA.},
\item{\emph{'upper'}:}{The part above the diagonal is used. the lower part is set to NA.}
}}

\item{diagonal}{Logical value.  Indicates if the diagonal values of the
matrices are kept ( = TRUE) or set to NA (= FALSE). Ignored if \code{type}
= \emph{'complete'}.}
}
\value{
Adjusted input matrix.
}
\description{
Adjust correlation matrix
}
