% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes-and-spectra.R
\name{all_color_palettes}
\alias{all_color_palettes}
\alias{all_color_spectra}
\title{Color palette names}
\usage{
all_color_palettes()

all_color_spectra()
}
\value{
A named list.
}
\description{
Returns all currently valid color panels or -spectra.
}
