% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes-and-spectra.R, R/deprecated.R
\name{all_color_palettes}
\alias{all_color_palettes}
\alias{all_color_spectra}
\alias{all_colorpanels}
\title{Color palette names}
\usage{
all_color_palettes()

all_color_spectra()

all_colorpanels()
}
\value{
A named list.
}
\description{
Returns all currently valid color palettes or -spectra.
}
\details{
\code{all_colorpanels()} is deprecated due to naming issues. Use
\code{all_color_palettes()} instead.
}
