% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{all_colorpanels}
\alias{all_colorpanels}
\title{Deprecated}
\usage{
all_colorpanels()
}
\description{
\code{all_colorpanels()} is deprecated due to naming issues. Use
\code{all_color_palettes()} instead.
}
