% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{argument_dummy}
\alias{argument_dummy}
\title{Argument dummy}
\usage{
argument_dummy(df, normalize, verbose)
}
\arguments{
\item{df}{A data.frame.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{across}{Character value. Specifies the discrete variable in the data.frame
across which the variables of interest are to be analyzed or displayed.}

\item{across.subset}{Character vector. The groups of interest that the grouping variable
denoted in \code{across} contains.

If there are more groups you are interested in than they are groups you are not interested
in specify those that you are not interested in prefixed with an \emph{'-'}.
Variables prefixed that way are discarded and the remaining are kept.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{all_color_palettes()} to obtain valid
input options.}

\item{clrsp}{Chracter value. Specfies the color spectrum to be used to represent
continuous values of numeric variables. Run \code{all_color_spectra()} to obtain
valid input options.}

\item{display.points}{Logical value. If set to TRUE points are used additionally
to display the results.}

\item{display.facets}{Logical value. If set to TRUE the plot is split via
\code{ggplot2::facet_wrap()} such that each variable gets it's own subplot.}

\item{key.name, key_name}{Character value or NULL. Denotes the variable that is used
to identify each observation uniquely.}

\item{key.prefix, key_prefix}{Character value. The string with which
the IDs in the newly constructed key variable are prefixed if \code{key_name} is
NULL.}

\item{pt.alpha}{Numeric value. Specifies the transparency of points.}

\item{pt.color}{Character value. Specifies the color with which all points are displayed .}

\item{pt.num}{Numeric value. Species the number of points (sample size)
to prevent overplotting.}

\item{pt.shape}{Numeric or character value. If numeric, the respective
shape is taken for all points. If character, the respective variable is
mapped to the shape-aesthetic.}

\item{pt.size}{Numeric values. Specifies size of points.}

\item{ref.group}{Character value. Specifies the reference group for the pairwise statistical test. Must
be among the groups the variable specified in \code{across} contains. If set to NULL the
first group found is taken.}

\item{relevel}{Logical value. If set to TRUE the input order of \code{across.subset} and
\code{variables} determines the order in which the groups of interest are displayed.}

\item{scales, nrow, ncol}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{test.pairwise}{Character value or NULL. If character, one of \emph{'t.test', 'wilcox.test'}.}

\item{test.groupwise}{Character value or NULL. If character, one of \emph{'anova', 'kruskal.test'}.}

\item{valid.classes}{Character vector to specify the classes the input
might have in order not to be discarded.}

\item{variables}{Character vector. Specifies the variables
of interest. If set to NULL all valid variables of the input data.frame are
considered (as long as the total number does not exceeds the limit).

If there are more variables you are interested in than they are variables
you are not interested in specify those that you are not interested in prefixed
with an \emph{'-'}. Variables prefixed that way are discarded and the remaining are kept.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Argument dummy
}
