% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame-handling.R
\name{arrange_rows}
\alias{arrange_rows}
\title{Arrange rows}
\usage{
arrange_rows(df, across = NULL, according.to, verbose)
}
\arguments{
\item{df}{A data.frame with at least one numeric variable.}

\item{according.to}{Character. Either \emph{'maxima} or \emph{'minima'}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The arranged data.frame.
}
\description{
Arranges the rows of a data.frame according to
the positions of their maxima or or minima. The earlier
a rows maximum/minimum appears the earlier the row will appear in the
returned data.frame.
}
