% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-wrapper-shinyjqui.R
\name{change_order_input}
\alias{change_order_input}
\title{Title}
\usage{
change_order_input(ns, id = "change_order", items, item_class = "default", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
