% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous-data-frame.R
\name{check_across_subset}
\alias{check_across_subset}
\title{Subset the across-variables}
\usage{
check_across_subset(
  df,
  across,
  across.subset,
  relevel = TRUE,
  fdb.fn = "warning"
)
}
\arguments{
\item{df}{A data.frame that contains the grouping variable specified in \code{across}.}

\item{across}{Character value. Denotes the discrete variable in the data.frame
across which the variables of interest are to be analyzed or displayed.}

\item{across.subset}{Character vector. The groups of interest that the grouping variable
denoted in \code{across} contains. If prefixed with an \emph{'-'} the groups are discarded.}

\item{relevel}{Logical value. If set to TRUE the input of \code{across.subset}
determines the new order in which the results are displayed.}
}
\value{
A filtered data.frame, informative messages or an error.
}
\description{
Checks across and across.subset input and if at least one
of the \code{across.subset} values exists if the letter is not NULL. It
then filters the input data.frame accordingly.
}
