% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous-data-frame.R
\name{check_across_subset}
\alias{check_across_subset}
\title{Subset the across-variables}
\usage{
check_across_subset(
  df,
  across,
  across.subset,
  relevel = TRUE,
  fdb.fn = "warning"
)
}
\arguments{
\item{df}{A data.frame that contains the grouping variable specified in \code{across}.}
}
\value{
A filtered data.frame, informative messages or an error.
}
\description{
Checks across and across.subset input and if at least one
of the \code{across.subset} values exists if the letter is not NULL. It
then filters the input data.frame accordingly.
}
