% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-check.R
\name{check_data_frame}
\alias{check_data_frame}
\title{Check data.frame validity}
\usage{
check_data_frame(df, var.class = list(), var.options = list(), ref = "df")
}
\arguments{
\item{df}{A data.frame.}

\item{var.class}{A named list. The names have to match the
variable names of the data.frame that are to be validated. The respective
elements specify the class the data.frame variable must have specified
as character strings.}

\item{ref}{Character value. Input reference for the error message.}
}
\value{
An informative error message or an invisible TRUE.
}
\description{
Checks whether the input data.frame contains variables
of certain classes and names.
}
\examples{
 # make sure that the input data.frame has
 # the numeric variables 'mpg' and 'cyl'.

 check_data_frame(df = mtcars,
                  var.class = list(mpg = "numeric",
                                   cyl = "numeric"))
}
