% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-check.R
\name{check_vector}
\alias{check_vector}
\title{Compare input to control input}
\usage{
check_vector(
  input,
  control,
  verbose = TRUE,
  ref.input = "input vector",
  ref.control = "control vector"
)
}
\arguments{
\item{input}{A vector of any kind.}

\item{control}{A vector of the same kind as \code{input}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{ref.input}{The reference character value for input.}

\item{ref.control}{The reference character value for control.}
}
\value{
An informative error message or an invisible TRUE.
}
\description{
Compare input to control input
}
