% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes-and-spectra.R
\name{color_vector}
\alias{color_vector}
\title{Get vector of colors}
\usage{
color_vector(clrp, names = NULL, clrp.adjust = NULL)
}
\arguments{
\item{clrp}{Character value. Denotes the color palette of interest. Run \code{all_colorpalettes()} to
obtain valid inputs.}

\item{names}{Character vector or NULL. Assigns names to the color vector (in the same order).}

\item{clrp.adjust}{Named character vector or NULL. If character, it adjusts the
colors that are used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}
}
\value{
Character vector.
}
\description{
Returns a vector of color codes of the respective palette
}
