% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes-and-spectra.R
\name{color_vector}
\alias{color_vector}
\title{Get vector of colors}
\usage{
color_vector(clrp)
}
\arguments{
\item{clrp}{Character value. Denotes the colorpanels of interest. Run \code{all_colorpanels()} to
obtain valid inputs.}
}
\value{
Character vector.
}
\description{
Returns a vector of colorcodes of the respective panel
}
