% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Clustering.R
\name{computeClusteringKmeans}
\alias{computeClusteringKmeans}
\alias{computeClusteringKmeans,Clustering-method}
\title{Compute cluster with kmeans}
\usage{
computeClusteringKmeans(object, ...)

\S4method{computeClusteringKmeans}{Clustering}(object, ks, methods_kmeans, verbose = TRUE)

\S4method{computeClusteringKmeans}{Clustering}(object, ks, methods_kmeans, verbose = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{ks}{Numeric vector. Denotes all options for k-clusters. Values <1 are discarded.
Is converted into an integer vector.}

\item{methods_kmeans}{Character vector. Specifies the kmeans methods
of interest. Use \code{validMethodsKmeans()} to obtain all valid input options.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object.
}
\description{
Compute cluster with method \emph{kmeans}.
}
