% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Correlation.R,
%   R/S4-Analysis.R
\name{computeCorrelation}
\alias{computeCorrelation}
\alias{computeCorrelation,Correlation-method}
\alias{computeCorrelation,Analysis-method}
\title{Compute correlation}
\usage{
computeCorrelation(object, ...)

\S4method{computeCorrelation}{Correlation}(
  object,
  across = NULL,
  methods_corr = "pearson",
  verbose = TRUE,
  ...
)

\S4method{computeCorrelation}{Analysis}(
  object,
  across = NULL,
  across_subset = NULL,
  methods_corr = "pearson",
  verbose = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{...}{Additional arguments given to \code{Hmisc::rcorr()}.}

\item{across}{Character vector or NULL. If character, the name of the grouping variables
that contain the group assignment across which analysis should
be conducted. If NULL, no grouping is done.}

\item{methods_corr}{Character vector. Specifies the correlation
methods of interest. Use \code{validMethodsCorrelation()} to obtain all valid
input options.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{across_subset}{Character vector or NULL. If NULL, all groups
are considered. If character, specifies the groups of the
grouping variable that are of interest. Prefixing group names with
\emph{'-'} excludes them.}
}
\value{
The input object.
}
\description{
Computes correlation matrix as well as corresponding
p-values.
}
\details{
If argument \code{across} is NULL. All numeric variables are simply
correlated.

If \code{across} is character: Prior to correlation, the data set is split into subsets for
each group a grouping variable contains. Then the variables are correlated separately
for each group. This is done for every grouping variable denoted in the input
vector of \code{across.}

Results can be retrieved or visualized by specifying the grouping
variable of interest in the \code{across} argument and the group(s) of interest
with the \code{across_subset} argument of the respective function.

Computation is conducted with \code{Hmisc::rcorr()}.
}
