% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Clustering.R
\name{computeDistanceMatrices}
\alias{computeDistanceMatrices}
\alias{computeDistanceMatrices,Clustering-method}
\title{Compute distance matrices}
\usage{
computeDistanceMatrices(object, ...)

\S4method{computeDistanceMatrices}{Clustering}(
  object,
  methods_dist = "euclidean",
  p = 2,
  force = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{methods_dist}{Character value. Specifies the distance methods
of interest. Use \code{validMethodsDist()} to obtain all valid input options.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object.
}
\description{
Computes distance matrices of the input data.
}
