% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Clustering.R,
%   R/S4-Analysis.R
\name{getClusterVarsPam}
\alias{getClusterVarsPam}
\alias{getClusterVarsPam,Clustering-method}
\alias{getClusterVarsPam,Analysis-method}
\title{Obtain cluster variables (PAM)}
\usage{
getClusterVarsPam(object, ...)

\S4method{getClusterVarsPam}{Clustering}(
  object,
  ks,
  methods_pam = "euclidean",
  prefix = "",
  naming = "{method_pam}_k{k}"
)

\S4method{getClusterVarsPam}{Analysis}(
  object,
  ks,
  methods_pam = "euclidean",
  prefix = "",
  naming = "{method_pam}_k{k}"
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{ks}{Numeric vector. Denotes all options for k-clusters. Values <1 are discarded.
Is converted into an integer vector.}
}
\value{
A data.frame.
}
\description{
Extracts a data.frame that contains grouping variables
according to clustering results.
}
