% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-outlierDetection.R
\name{getDf}
\alias{getDf}
\alias{getDf,OutlierDetection-method}
\title{Obtain object data}
\usage{
getDf(object)

\S4method{getDf}{OutlierDetection}(object)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}
}
\value{
A data.frame with subclass \code{tibble}.
}
\description{
Extracts the objects data as a data.frame.
}
