% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-Outlier.R, R/S4-Analysis.R
\name{getOutlierResults}
\alias{getOutlierResults}
\alias{getOutlierResults,OutlierDetection-method}
\alias{getOutlierResults,Analysis-method}
\title{Obtain outlier results}
\usage{
getOutlierResults(object, ...)

\S4method{getOutlierResults}{OutlierDetection}(object, method = "IQR", across = NULL, verbose = TRUE)

\S4method{getOutlierResults}{Analysis}(object, method = "IQR", across = NULL, verbose = TRUE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{method}{Character value. The outlier detection method. Valid
input options are \emph{'IQR'} and \emph{'Mahalanobis'}.}

\item{across}{Character value or NULL. If character, denotes
the grouping variable of interest.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The return value depends on the method chosen:

\itemize{
\item{\code{method} = \emph{'IQR'}:}{
A named list in which each slot is a vector of IDs which were identified
as outliers within a numeric variable denoted by the name of the list's slot.
}
\item{\code{method} = \emph{'Mahalanobis'}:}{
A data.frame  with column names \emph{<key.name>}, \emph{mahal} and \emph{pval}.
}
}

If \code{across} is a character value the returned value is a list
named according to the groups of the grouping variable \code{across} each containing
output as described above.
}
\description{
Extracts outlier results from the object.
}
