% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-outlierDetection.R
\name{getResults}
\alias{getResults}
\alias{getResults,OutlierDetection-method}
\alias{getResults,OutlierDetectionMethod-method}
\title{Obtain analysis results}
\usage{
getResults(object, ...)

\S4method{getResults}{OutlierDetection}(object, method)

\S4method{getResults}{OutlierDetectionMethod}(object, across)
}
\arguments{
\item{...}{}

\item{across}{Character value. Specifies the discrete variable in the data.frame
across which the variables of interest are to be analyzed or displayed.}
}
\value{
Depends on the objects class.
}
\description{
Generic extractor for results mainly used for
programming purpose as it provides informative error
messages if the requested content is missing.
}
