% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-AnalysisAspect.R
\name{getScaledDf}
\alias{getScaledDf}
\alias{getScaledDf,AnalysisAspect-method}
\title{Obtain scaled data}
\usage{
getScaledDf(object, ...)

\S4method{getScaledDf}{AnalysisAspect}(object, grouping = FALSE, logical = FALSE)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{grouping}{Logical value. Indicate if the respective variable
types should be part of the output data.frame.}

\item{logical}{Logical value. Indicate if the respective variable
types should be part of the output data.frame.}
}
\value{
A data.frame.
}
\description{
Extracts the scaled data in a data.frame. Grouping and
logical variables can be joined
}
