% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Clustering.R,
%   R/S4-ClusteringPam.R
\name{getSilWidthsDf}
\alias{getSilWidthsDf}
\alias{getSilWidthsDf,Clustering-method}
\alias{getSilWidthsDf,ClusteringPam-method}
\title{Obtain sil-width data}
\usage{
getSilWidthsDf(object, ...)

\S4method{getSilWidthsDf}{Clustering}(object, ks, method_pam = "euclidean")

\S4method{getSilWidthsDf}{ClusteringPam}(object, ks, method_pam = "euclidean")
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{ks}{Numeric vector. Denotes all options for k-clusters. Values <1 are discarded.
Is converted into an integer vector.}

\item{method_pam}{Character value. Specifies the pam-method of interest.
Use \code{validMethodsPam()} to obtain all valid input options.}
}
\value{
Data.frame.
}
\description{
Extracts a data.frame that contains information
about the silhouette width of every observation in different
clustering results with method \emph{pam}.
}
