% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-check.R
\name{give_feedback}
\alias{give_feedback}
\title{Print feedback in console}
\usage{
give_feedback(fdb.fn = c("stop", "warning", "message"), msg = NULL)
}
\arguments{
\item{fdb.fn}{Character value. Determines the function to call if a feedback
message needs to be given. One of \emph{'stop', 'warning'} or \emph{'message'}.}

\item{msg}{Character value or glue. The message to be printed in the console.}
}
\value{

}
\description{
Helper that gives feedback with a function of choice.
}
