% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust-wrapper.R
\name{initiate_hclust_object}
\alias{initiate_hclust_object}
\title{Title}
\usage{
initiate_hclust_object(
  hclust_data = NULL,
  key_name = NULL,
  hclust.data = NULL,
  key.name = NULL,
  scale = TRUE,
  default.dist = "euclidean",
  default.aggl = "complete",
  default.dir = "conv-hcl-obj.RDS",
  verbose = TRUE
)
}
\arguments{
\item{hclust_data}{}
}
\value{

}
\description{
Title
}
