% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-check.R
\name{is_value}
\alias{is_value}
\alias{is_vec}
\title{One dimensional input check}
\usage{
is_value(x, mode, ref = "x")

is_vec(x, mode, ref = "x")
}
\arguments{
\item{x}{Input vector.}

\item{mode}{Character value. The type of which the input must be.}

\item{ref}{Character value. Input reference for the error message.}
}
\value{
An invisible TRUE or an informative error message.
}
\description{
Checks if input fits the requirements and stops the
function if not.
}
