% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-Correlation.R
\name{melt_rcorr}
\alias{melt_rcorr}
\title{Melt \code{rcorr}}
\usage{
melt_rcorr(rcorr_obj, type = "complete", diagonal = TRUE, distinct = FALSE)
}
\arguments{
\item{rcorr_obj}{An object of class \code{rcorr}.}

\item{type}{Character value. Denotes how the underlying correlation matrix is
handled. Three options:

\itemize{
\item{\emph{'complete'}:}{ The matrix stays as is.},
\item{\emph{'lower'}:}{ The part below the diagonal is used. The upper part is set to NA.},
\item{\emph{'upper'}:}{The part above the diagonal is used. the lower part is set to NA.}
}}

\item{diagonal}{Logical value.  Indicates if the diagonal values of the
matrices are kept ( = TRUE) or set to NA (= FALSE). Ignored if \code{type}
= \emph{'complete'}.}

\item{distinct}{Logical value. If TRUE only one observation per variable pair remains
in the output data.frame.}
}
\value{
A data.frame with the following columns:

\itemize{
\item{\emph{var1}:}{ Factor. First variable of the correlated variable pair.}
\item{\emph{var2}:}{ Factor. Second variable of the correlated variable pair.},
\item{\emph{corr}:}{ Numeric. The correlation vaule.},
\item{\emph{pval}:}{ Numeric. The corresponding p-value.},
}
}
\description{
Melts object of class \code{rcorr} to a data.frame.
}
