% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{normalize_dummy}
\alias{normalize_dummy}
\title{normalize_dummy}
\usage{
normalize_dummy()
}
\arguments{
\item{normalize}{Logical. If set to TRUE numeric values will be scaled down to
values between one and zero.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
normalize_dummy

ggplot2_dummy
}
