% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math-related-functions.R
\name{one_peak}
\alias{one_peak}
\alias{two_peaks}
\alias{gradient}
\alias{log}
\title{Fit a curve}
\usage{
one_peak(input)

two_peaks(input)

gradient(input)

log(input)
}
\arguments{
\item{input}{Numeric vector}
}
\description{
Functions that return the numeric values of the respective
mathematic function they are constructed of.
}
