% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math-related-functions.R
\name{one_peak}
\alias{one_peak}
\alias{two_peaks}
\alias{linear}
\alias{sinus}
\alias{gradient}
\alias{log}
\alias{early_peak}
\alias{late_peak}
\alias{abrupt_ascending}
\alias{abrupt_descending}
\alias{immediate_ascending}
\alias{immediate_descending}
\title{Fit a curve}
\usage{
one_peak(input)

two_peaks(input)

linear(input)

sinus(input)

gradient(input)

log(input)

early_peak(input)

late_peak(input)

abrupt_ascending(input)

abrupt_descending(input)

immediate_ascending(input)

immediate_descending(input)
}
\arguments{
\item{input}{Numeric vector.}
}
\description{
Functions that return the numeric values of the respective
mathematic function they are constructed of. Length and range of the
return vector are equal to those of the input vector.
}
