% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-module-descriptive-statistics.R
\name{plot_descriptive_statistics_interactive}
\alias{plot_descriptive_statistics_interactive}
\title{Opens interactive application}
\usage{
plot_descriptive_statistics_interactive(df, n.across.subset = 10)
}
\description{
Opens interactive application
}
