% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dot-plot.R
\name{plot_dot_plot_2d}
\alias{plot_dot_plot_2d}
\title{Plot 2d dot plot}
\usage{
plot_dot_plot_2d(
  df,
  x,
  y,
  alpha.by = NULL,
  alpha.trans = "identity",
  color.by = NULL,
  color.trans = "identity",
  shape.by = NULL,
  size.by = NULL,
  size.trans = "reverse",
  pt.alpha = 0.9,
  pt.color = "black",
  pt.clrp = "default",
  pt.clrsp = "plasma",
  pt.shape = 19,
  pt.size = 3,
  transform.with = NULL,
  arrange.y = FALSE,
  arrange.by = NULL,
  reverse.all = FALSE,
  reverse.within = FALSE,
  ...
)
}
\arguments{
\item{df}{A data.frame.}

\item{x}{Character value. Variable to be plotted on the respective axis.}

\item{y}{Character value. Variable to be plotted on the respective axis.}

\item{alpha.by}{Character
value or NULL. If character, specifies the variable that is mapped to the respective
aesthetic of the plot.}

\item{alpha.trans}{Character value. Name of the transformation method applied to the continuous scale. Use
\code{validScaleTransformations()} to obtain all valid input options. Use \emph{'identity'}
for no transformation.}

\item{color.by}{Character value or NULL. If character, specifies the
variable that is displayed by color.}

\item{color.trans}{Character value. Name of the transformation method applied to the continuous scale. Use
\code{validScaleTransformations()} to obtain all valid input options. Use \emph{'identity'}
for no transformation.}

\item{shape.by}{Character
value or NULL. If character, specifies the variable that is mapped to the respective
aesthetic of the plot.}

\item{size.by}{Character
value or NULL. If character, specifies the variable that is mapped to the respective
aesthetic of the plot.}

\item{size.trans}{Character value. Name of the transformation method applied to the continuous scale. Use
\code{validScaleTransformations()} to obtain all valid input options. Use \emph{'identity'}
for no transformation.}

\item{pt.alpha}{Numeric value. Specifies the transparency of points.}

\item{pt.color}{Character value. Specifies the color with which all points are displayed .}

\item{pt.clrsp}{Character value. Specifies the color spectrum to be used to represent
continuous values of numeric variables. Run \code{validColorSpectra()} to obtain
valid input options.}

\item{pt.shape}{Numeric or character value. If numeric, the respective
shape is taken for all points. If character, the respective variable is
mapped to the shape-aesthetic.}

\item{pt.size}{Numeric values. Specifies size of points.}

\item{transform.with}{List or NULL. If list, can be used to transform continuous variables before plotting.
Names of the slots of the provided list refer to the variables. The content of the slot refers to the transforming functions.
Slot content can either be a character vector of function names. Use \code{validVarTransformations()} to obtain all valid character value inputs.
Or it can be a list of functions (and function names).}

\item{arrange.y}{Logical. If TRUE the labels of \code{y} are arranged by
their appearances in the groups of \code{x}.}

\item{arrange.by}{Character value or NULL. If character, denotes the
numeric variable by which the labels of \code{y} are arranged within the
groups of \code{x}.}

\item{reverse.all}{Logical. If TRUE labels are displayed from bottom to top.
If FALSE, labels are displayed from top to bottom.}

\item{reverse.within}{Logical. If TRUE, labels are arranged within the
groups in a descending manner. If FALSE, labels are arranged within the
groups in an ascending manner. (regarding the numeric variables of \code{arrange.by}).}
}
\value{
A ggplot.
}
\description{
Plots a dot plot of two categorical variables. Numeric
values can be displayed via alpha, color and size.
}
