% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-scatter.R
\name{plot_scatterplot}
\alias{plot_scatterplot}
\title{Title}
\usage{
plot_scatterplot(
  df,
  x,
  y,
  across = NULL,
  across.subset = NULL,
  relevel = TRUE,
  ncol = NULL,
  nrow = NULL,
  scales = "fixed",
  space = "fixed",
  clr.by = NULL,
  pt.alpha = 0.9,
  pt.clr = "black",
  pt.fill = "black",
  pt.shape = 21,
  pt.size = 1.5,
  display.smooth = FALSE,
  smooth.alpha = 0.9,
  smooth.clr = "blue",
  smooth.method = "lm",
  smooth.se = FALSE,
  display.corr = FALSE,
  corr.method = "pearson",
  corr.p.min = 5e-05,
  corr.pos.x = NULL,
  corr.pos.y = NULL,
  corr.text.sep = "\\n",
  corr.text.size = 1
)
}
\arguments{
\item{smooth.se}{}
}
\value{

}
\description{
Title
}
