% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-scatter.R
\name{plot_scatterplot}
\alias{plot_scatterplot}
\title{Title}
\usage{
plot_scatterplot(
  df,
  x,
  y,
  across = NULL,
  across.subset = NULL,
  relevel = TRUE,
  ncol = NULL,
  nrow = NULL,
  scales = "fixed",
  space = "fixed",
  pt.alpha = 0.9,
  pt.clr = NA,
  pt.color = "black",
  pt.fill = "black",
  pt.shape = 19,
  pt.size = 1.5,
  alpha.by = NULL,
  color.aes = "color",
  color.by = NULL,
  color.trans = "identity",
  order.by = NULL,
  order.desc = FALSE,
  shape.by = NULL,
  size.by = NULL,
  clrp = "milo",
  clrp.adjust = NULL,
  clrsp = "inferno",
  display.smooth = FALSE,
  smooth.alpha = 0.9,
  smooth.color = "blue",
  smooth.method = "lm",
  smooth.se = FALSE,
  smooth.size = 1,
  smooth.type = "solid",
  display.corr = FALSE,
  corr.method = "pearson",
  corr.p.min = 5e-05,
  corr.pos.x = NULL,
  corr.pos.y = NULL,
  corr.text.sep = "\\n",
  corr.text.size = 1,
  transform.with = NULL,
  clr.aes = NA,
  clr.by = NA,
  pt.clrp = NA,
  ...
)
}
\arguments{
\item{df}{A data.frame.}

\item{x, y}{Character value. Variable to be plotted on the respective axis.}

\item{across}{Character value. Specifies the discrete variable in the data.frame
across which the variables of interest are to be analyzed or displayed.}

\item{across.subset}{Character vector. The groups of interest that the grouping variable
denoted in \code{across} contains.

If there are more groups you are interested in than they are groups you are not interested
in specify those that you are not interested in prefixed with an \emph{'-'}.
Variables prefixed that way are discarded and the remaining are kept.}

\item{relevel}{Logical value. If set to TRUE the input order of \code{across.subset} and
\code{variables} determines the order in which the groups of interest are displayed.}

\item{scales, nrow, ncol}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{order.desc}{Logical value. If set to TRUE the way the points are ordered
is reversed.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}
}
\value{
A ggplot.
}
\description{
Title
}
