% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-statistics-single.R
\name{plot_statistics}
\alias{plot_statistics}
\title{Plot distribution and results of statistical tests}
\usage{
plot_statistics(df, plot_type = "violin", ...)
}
\arguments{
\item{df}{A data.frame.}

\item{plot_type}{Character value. Denotes the function to call. Must
be one of \emph{'violin', 'boxplot', 'density', 'histogram', 'ridgeplot'}.}

\item{...}{Arguments given to the called function.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Plot distribution and results of statistical tests
}
