% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-statistics.R
\name{plot_statistics}
\alias{plot_statistics}
\title{Opens interactive application}
\usage{
plot_statistics(df, n.across.subset = 10)
}
\arguments{
\item{df}{A data.frame.}
}
\description{
Opens interactive application
}
