% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Clustering.R
\name{plotDendrogram}
\alias{plotDendrogram}
\alias{plotDendrogram,Clustering-method}
\title{Plot dendrograms}
\usage{
plotDendrogram(object, ...)

\S4method{plotDendrogram}{Clustering}(
  object,
  methods_dist = "euclidean",
  methods_aggl = "ward.D",
  k = NULL,
  h = NULL,
  type = "rectangle",
  facet_with = "grid",
  direction = "bt",
  branch_color = "black",
  branch_size = 1,
  display_labels = FALSE,
  labels_angle = 90,
  labels_hjust = 0,
  labels_nudge = 0.01,
  labels_size = 3,
  labels_vjust = 0.5,
  display_legend = TRUE,
  display_title = FALSE,
  clrp = "milo",
  clrp_adjust = NULL,
  simple = FALSE,
  nrow = NULL,
  ncol = NULL,
  ...
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{methods_dist}{Character value. Specifies the distance methods
of interest. Use \code{validMethodsDist()} to obtain all valid input options.}

\item{methods_aggl}{Character vector. Specifies the agglomerative
methods of interest. Use \code{validMethodsAggl()} to obtain all valid input options.}

\item{k}{Numeric value. Denotes the number of clusters. Must be bigger than 1. Must
not be bigger than the number of observations of the data set.}

\item{h}{Numeric value. Denotes the height at which the dendrogram is cut.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector. Can be used to adjust colors
with which groups are displayed. Names indicate groups, values
indicate the colors with which specific groups are supposed to be displayed.}

\item{nrow}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{ncol}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}
}
\value{
A ggplot or a base plot.
}
\description{
Plots a dendrogram with either ggplot or base plot.
}
