% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Clustering.R,
%   R/S4-Analysis.R
\name{plotScreeplot}
\alias{plotScreeplot}
\alias{plotScreeplot,Clustering-method}
\alias{plotScreeplot,Analysis-method}
\title{Plot a screeplot}
\usage{
plotScreeplot(object, ...)

\S4method{plotScreeplot}{Clustering}(
  object,
  methods_kmeans,
  ks = NULL,
  color = "steelblue",
  display_cols = TRUE,
  display_line = TRUE,
  display_points = TRUE
)

\S4method{plotScreeplot}{Analysis}(
  object,
  method_kmeans = "Hartigan-Wong",
  ks = NULL,
  color = "steelblue",
  display_cols = TRUE,
  display_points = TRUE,
  display_line = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{methods_kmeans}{Character vector. Specifies the kmeans methods
of interest. Use \code{validMethodsKmeans()} to obtain all valid input options.}

\item{ks}{Numeric vector. Denotes all options for k-clusters. Values <1 are discarded.
Is converted into an integer vector.}

\item{method_kmeans}{Character value. Specifies the kmeans method
of interest. Use \code{validMethodsKmeans()} to obtain all valid input options.}
}
\value{
A ggplot.
}
\description{
Plots quality of clustering results
suggested by kmeans with a screeplot.
}
\details{
If \code{ks} and \code{methods_kmeans} are specified only
the resulting combinations are included. Else all found combinations
are included.
}
