% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-Clustering.R
\name{plotSilWidths}
\alias{plotSilWidths}
\alias{plotSilWidths,Clustering-method}
\title{Plot sil-width data}
\usage{
plotSilWidths(object, ...)

\S4method{plotSilWidths}{Clustering}(
  object,
  ks,
  method_pam = "euclidean",
  clrp = "milo",
  ncol = NULL,
  nrow = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{ks}{Numeric vector. Denotes all options for k-clusters. Values <1 are discarded.
Is converted into an integer vector.}

\item{method_pam}{Character value. Specifies the pam-method of interest.
Use \code{validMethodsPam()} to obtain all valid input options.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{ncol}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{nrow}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A ggplot.
}
\description{
Plots information about the silhouette width of
every observation in different clustering results with method \emph{pam}.
}
