% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R, R/S4-DimRed.R
\name{plotUMAP}
\alias{plotUMAP}
\alias{plotUMAP,DimRed-method}
\title{Plot UMAP}
\usage{
plotUMAP(object, ...)

\S4method{plotUMAP}{DimRed}(
  object,
  alpha_by = NULL,
  color_by = NULL,
  shape_by = NULL,
  size_by = NULL,
  pt_alpha = 0.9,
  pt_color = "black",
  pt_fill = "black",
  pt_shape = 19,
  pt_size = 1,
  color_aes = "color",
  clrp = "milo",
  clrp_adjust = NULL,
  clrsp = "inferno",
  ...
)
}
\arguments{
\item{object}{Any object for whose class a method has been defined.}

\item{alpha_by}{Character
value or NULL. If character, specifies the variable that is mapped to the respective
aesthetic of the plot.}

\item{color_by}{Character
value or NULL. If character, specifies the variable that is mapped to the respective
aesthetic of the plot.}

\item{shape_by}{Character
value or NULL. If character, specifies the variable that is mapped to the respective
aesthetic of the plot.}

\item{size_by}{Character
value or NULL. If character, specifies the variable that is mapped to the respective
aesthetic of the plot.}

\item{pt_alpha}{Numeric value. Specifies the transparency of points.}

\item{pt_color}{Character value. Specifies the color with which all points are displayed .}

\item{pt_shape}{Numeric or character value. If numeric, the respective
shape is taken for all points. If character, the respective variable is
mapped to the shape-aesthetic.}

\item{color_aes}{Either \emph{'color'} or \emph{'fill'}. Specifies the
aesthetic that is used to visualize the variable that is specified in \code{color_by}.

Ignored if \code{color_by} is NULL.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{all_color_palettes()} to obtain valid
input options.}

\item{clrsp}{Chracter value. Specfies the color spectrum to be used to represent
continuous values of numeric variables. Run \code{all_color_spectra()} to obtain
valid input options.}
}
\value{
A ggplot.
}
\description{
Plots UMAP results in a scatterplot.
}
