% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_infos.R
\name{print_gene_info}
\alias{print_gene_info}
\title{Print gene information}
\usage{
print_gene_info(genes, check = TRUE)
}
\arguments{
\item{genes}{Character vector of gene names.}
}
\value{
Invisible TRUE. Texts are immediately printed using \code{base::writeLines()}.
}
\description{
Prints human readable summary texts of genes in the console.
}
