% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes-and-spectra.R
\docType{data}
\name{sequential_single_hue}
\alias{sequential_single_hue}
\alias{sequential_multi_hue}
\alias{diverging}
\title{Color spectra names}
\format{
An object of class \code{character} of length 11.

An object of class \code{character} of length 61.

An object of class \code{character} of length 18.
}
\usage{
sequential_single_hue

sequential_multi_hue

diverging
}
\description{
Collection of color-spectra names.
}
\keyword{datasets}
