% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generics.R
\name{set_data_hlpr}
\alias{set_data_hlpr}
\title{Set data helper}
\usage{
set_data_hlpr(
  object,
  data,
  key.name,
  key.prefix,
  slot.key.name = "key_name",
  slot.data = "data",
  verbose = TRUE
)
}
\arguments{
\item{key.name}{Character value or NULL. Denotes the variable that is used
to identify each observation uniquely.}

\item{key.prefix}{Character value. The string with which
the IDs in the newly constructed key variable are prefixed if \code{key_name} is
NULL.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object.
}
\description{
Helper for generic function \code{setData()}.
}
