% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-wrapper-shiny.R
\name{slider_input_alpha}
\alias{slider_input_alpha}
\alias{slider_input_size}
\title{Slider inputs}
\usage{
slider_input_alpha(ns, id = "pt.alpha")

slider_input_size(
  ns,
  id = "pt.size",
  min = 0.1,
  max = 5,
  step = 0.1,
  value = 1.2
)
}
\description{
Slider inputs
}
