% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_infos.R
\name{synonyms_to_hgnc}
\alias{synonyms_to_hgnc}
\title{Map gene synonyms to HGNC symbols}
\usage{
synonyms_to_hgnc(input, verbose = TRUE, ...)
}
\arguments{
\item{input}{Character vector of possible gene synonyms.}

\item{verbose}{Logical. If TRUE, informative message are printed in
the console about which elements of \code{input} could be matched
to which genes and which could not be mapped.}
}
\value{
Character vector.
}
\description{
Maps gene synonyms to their official HGNC symbol.
}
\details{
Elements of \code{input} that are, in fact, synonyms are
replaced by their official HGNC symbol. Elements that are already
HGNC symbols stay as they are. Elements that are neither are silently
dropped.
}
\examples{

synonyms_to_hgnc(input = c("GFAP", "M33", "XXX"), verbose = T)


}
