% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-AnalysisAspect.R, R/S4-Outlier.R,
%   R/S4-DimRed.R, R/S4-Clustering.R, R/S4-Correlation.R,
%   R/S4-ClusteringHclust.R, R/S4-ClusteringKmeans.R, R/S4-ClusteringPam.R,
%   R/miscellaneous.R
\name{validAnalysisAspects}
\alias{validAnalysisAspects}
\alias{validMethodsOutlierDetection}
\alias{validMethodsDimRed}
\alias{validMethodsClustering}
\alias{validMethodsCorrelation}
\alias{validShapesCorrelation}
\alias{validTypesCorrelation}
\alias{validMethodsAggl}
\alias{validMethodsDist}
\alias{validMethodsKmeans}
\alias{validMethodsPam}
\alias{validInput}
\alias{validCurves}
\alias{validLineTypes}
\title{Obtain valid input options}
\usage{
validAnalysisAspects()

validMethodsOutlierDetection()

validMethodsDimRed()

validMethodsClustering()

validMethodsCorrelation()

validShapesCorrelation()

validTypesCorrelation()

validMethodsAggl()

validMethodsDist()

validMethodsKmeans()

validMethodsPam()

validInput()

validCurves()

validLineTypes()
}
\value{
Character vector.
}
\description{
These functions return all valid input options for
specific arguments.
}
