% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-classes.R
\name{variable_classes2}
\alias{variable_classes2}
\title{Obtain classes with names}
\usage{
variable_classes2(data, simplify = TRUE)
}
\arguments{
\item{data}{A data.frame or a list.}

\item{simplify}{Logical. If set to TRUE the return value is turned into
a character vector.}
}
\value{
A named character vector.
\itemize{
\item{ \strong{Vector-elements:} The class of every variable of \code{data}}.
\item{ \strong{Vector-names:} The name of every variable of \code{data}}.
}
}
\description{
Get information about the inputs
variable names and classes.
}
\examples{
variable_classes2(data = mtcars)

}
