% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-leverage.R
\name{vrename}
\alias{vrename}
\alias{vrename_with}
\title{Rename vector in tidyverse style}
\usage{
vrename(input, ..., safely = TRUE)

vrename_with(input, .fn, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Rename vector in tidyverse style
}
